Crate rusoto_lightsail

Amazon Lightsail is the easiest way to get started with AWS for developers who just need virtual private servers. Lightsail includes everything you need to launch your project quickly - a virtual machine, a managed database, SSD-based storage, data transfer, DNS management, and a static IP - for a low, predictable price. You manage those Lightsail servers through the Lightsail console or by using the API or command-line interface (CLI).

For more information about Lightsail concepts and tasks, see the Lightsail Dev Guide.

To use the Lightsail API or the CLI, you will need to use AWS Identity and Access Management (IAM) to generate access keys. For details about how to set this up, see the Lightsail Dev Guide.

If you're using the service, you're probably looking for LightsailClient and Lightsail.

Structs

AllocateStaticIpRequest

AllocateStaticIpResult

AttachDiskRequest

AttachDiskResult

AttachInstancesToLoadBalancerRequest

AttachInstancesToLoadBalancerResult

AttachLoadBalancerTlsCertificateRequest

AttachLoadBalancerTlsCertificateResult

AttachStaticIpRequest

AttachStaticIpResult

AvailabilityZone
Describes an Availability Zone.

Blueprint
Describes a blueprint (a virtual private server image).

Bundle
Describes a bundle, which is a set of specs describing your virtual private server (or instance).

CloseInstancePublicPortsRequest

CloseInstancePublicPortsResult

CloudFormationStackRecord
Describes a CloudFormation stack record created as a result of the create cloud formation stack operation.

A CloudFormation stack record provides information about the AWS CloudFormation stack used to create a new Amazon Elastic Compute Cloud instance from an exported Lightsail instance snapshot.

CloudFormationStackRecordSourceInfo
Describes the source of a CloudFormation stack record (i.e., the export snapshot record).

CopySnapshotRequest

CopySnapshotResult

CreateCloudFormationStackRequest

CreateCloudFormationStackResult

CreateDiskFromSnapshotRequest

CreateDiskFromSnapshotResult

CreateDiskRequest

CreateDiskResult

CreateDiskSnapshotRequest

CreateDiskSnapshotResult

CreateDomainEntryRequest

CreateDomainEntryResult

CreateDomainRequest

CreateDomainResult

CreateInstanceSnapshotRequest

CreateInstanceSnapshotResult

CreateInstancesFromSnapshotRequest

CreateInstancesFromSnapshotResult

CreateInstancesRequest

CreateInstancesResult

CreateKeyPairRequest

CreateKeyPairResult

CreateLoadBalancerRequest

CreateLoadBalancerResult

CreateLoadBalancerTlsCertificateRequest

CreateLoadBalancerTlsCertificateResult

CreateRelationalDatabaseFromSnapshotRequest

CreateRelationalDatabaseFromSnapshotResult

CreateRelationalDatabaseRequest

CreateRelationalDatabaseResult

CreateRelationalDatabaseSnapshotRequest

CreateRelationalDatabaseSnapshotResult

DeleteDiskRequest

DeleteDiskResult

DeleteDiskSnapshotRequest

DeleteDiskSnapshotResult

DeleteDomainEntryRequest

DeleteDomainEntryResult

DeleteDomainRequest

DeleteDomainResult

DeleteInstanceRequest

DeleteInstanceResult

DeleteInstanceSnapshotRequest

DeleteInstanceSnapshotResult

DeleteKeyPairRequest

DeleteKeyPairResult

DeleteKnownHostKeysRequest

DeleteKnownHostKeysResult

DeleteLoadBalancerRequest

DeleteLoadBalancerResult

DeleteLoadBalancerTlsCertificateRequest

DeleteLoadBalancerTlsCertificateResult

DeleteRelationalDatabaseRequest

DeleteRelationalDatabaseResult

DeleteRelationalDatabaseSnapshotRequest

DeleteRelationalDatabaseSnapshotResult

DestinationInfo
Describes the destination of a record.

DetachDiskRequest

DetachDiskResult

DetachInstancesFromLoadBalancerRequest

DetachInstancesFromLoadBalancerResult

DetachStaticIpRequest

DetachStaticIpResult

Disk
Describes a system disk or an block storage disk.

DiskInfo
Describes a disk.

DiskMap
Describes a block storage disk mapping.

DiskSnapshot
Describes a block storage disk snapshot.

DiskSnapshotInfo
Describes a disk snapshot.

Domain
Describes a domain where you are storing recordsets in Lightsail.

DomainEntry
Describes a domain recordset entry.

DownloadDefaultKeyPairRequest

DownloadDefaultKeyPairResult

ExportSnapshotRecord
Describes an export snapshot record.

ExportSnapshotRecordSourceInfo
Describes the source of an export snapshot record.

ExportSnapshotRequest

ExportSnapshotResult

GetActiveNamesRequest

GetActiveNamesResult

GetBlueprintsRequest

GetBlueprintsResult

GetBundlesRequest

GetBundlesResult

GetCloudFormationStackRecordsRequest

GetCloudFormationStackRecordsResult

GetDiskRequest

GetDiskResult

GetDiskSnapshotRequest

GetDiskSnapshotResult

GetDiskSnapshotsRequest

GetDiskSnapshotsResult

GetDisksRequest

GetDisksResult

GetDomainRequest

GetDomainResult

GetDomainsRequest

GetDomainsResult

GetExportSnapshotRecordsRequest

GetExportSnapshotRecordsResult

GetInstanceAccessDetailsRequest

GetInstanceAccessDetailsResult

GetInstanceMetricDataRequest

GetInstanceMetricDataResult

GetInstancePortStatesRequest

GetInstancePortStatesResult

GetInstanceRequest

GetInstanceResult

GetInstanceSnapshotRequest

GetInstanceSnapshotResult

GetInstanceSnapshotsRequest

GetInstanceSnapshotsResult

GetInstanceStateRequest

GetInstanceStateResult

GetInstancesRequest

GetInstancesResult

GetKeyPairRequest

GetKeyPairResult

GetKeyPairsRequest

GetKeyPairsResult

GetLoadBalancerMetricDataRequest

GetLoadBalancerMetricDataResult

GetLoadBalancerRequest

GetLoadBalancerResult

GetLoadBalancerTlsCertificatesRequest

GetLoadBalancerTlsCertificatesResult

GetLoadBalancersRequest

GetLoadBalancersResult

GetOperationRequest

GetOperationResult

GetOperationsForResourceRequest

GetOperationsForResourceResult

GetOperationsRequest

GetOperationsResult

GetRegionsRequest

GetRegionsResult

GetRelationalDatabaseBlueprintsRequest

GetRelationalDatabaseBlueprintsResult

GetRelationalDatabaseBundlesRequest

GetRelationalDatabaseBundlesResult

GetRelationalDatabaseEventsRequest

GetRelationalDatabaseEventsResult

GetRelationalDatabaseLogEventsRequest

GetRelationalDatabaseLogEventsResult

GetRelationalDatabaseLogStreamsRequest

GetRelationalDatabaseLogStreamsResult

GetRelationalDatabaseMasterUserPasswordRequest

GetRelationalDatabaseMasterUserPasswordResult

GetRelationalDatabaseMetricDataRequest

GetRelationalDatabaseMetricDataResult

GetRelationalDatabaseParametersRequest

GetRelationalDatabaseParametersResult

GetRelationalDatabaseRequest

GetRelationalDatabaseResult

GetRelationalDatabaseSnapshotRequest

GetRelationalDatabaseSnapshotResult

GetRelationalDatabaseSnapshotsRequest

GetRelationalDatabaseSnapshotsResult

GetRelationalDatabasesRequest

GetRelationalDatabasesResult

GetStaticIpRequest

GetStaticIpResult

GetStaticIpsRequest

GetStaticIpsResult

HostKeyAttributes
Describes the public SSH host keys or the RDP certificate.

ImportKeyPairRequest

ImportKeyPairResult

Instance
Describes an instance (a virtual private server).

InstanceAccessDetails
The parameters for gaining temporary access to one of your Amazon Lightsail instances.

InstanceEntry
Describes the Amazon Elastic Compute Cloud instance and related resources to be created using the create cloud formation stack operation.

InstanceHardware
Describes the hardware for the instance.

InstanceHealthSummary
Describes information about the health of the instance.

InstanceNetworking
Describes monthly data transfer rates and port information for an instance.

InstancePortInfo
Describes information about the instance ports.

InstancePortState
Describes the port state.

InstanceSnapshot
Describes the snapshot of the virtual private server, or instance.

InstanceSnapshotInfo
Describes an instance snapshot.

InstanceState
Describes the virtual private server (or instance) status.

IsVpcPeeredRequest

IsVpcPeeredResult

KeyPair
Describes the SSH key pair.

LightsailClient
A client for the Amazon Lightsail API.

LoadBalancer
Describes the Lightsail load balancer.

LoadBalancerTlsCertificate
Describes a load balancer SSL/TLS certificate.

TLS is just an updated, more secure version of Secure Socket Layer (SSL).

LoadBalancerTlsCertificateDomainValidationOption
Contains information about the domain names on an SSL/TLS certificate that you will use to validate domain ownership.

LoadBalancerTlsCertificateDomainValidationRecord
Describes the validation record of each domain name in the SSL/TLS certificate.

LoadBalancerTlsCertificateRenewalSummary
Contains information about the status of Lightsail's managed renewal for the certificate.

LoadBalancerTlsCertificateSummary
Provides a summary of SSL/TLS certificate metadata.

LogEvent
Describes a database log event.

MetricDatapoint
Describes the metric data point.

MonthlyTransfer
Describes the monthly data transfer in and out of your virtual private server (or instance).

OpenInstancePublicPortsRequest

OpenInstancePublicPortsResult

Operation
Describes the API operation.

PasswordData
The password data for the Windows Server-based instance, including the ciphertext and the key pair name.

PeerVpcRequest

PeerVpcResult

PendingMaintenanceAction
Describes a pending database maintenance action.

PendingModifiedRelationalDatabaseValues
Describes a pending database value modification.

PortInfo
Describes information about the ports on your virtual private server (or instance).

PutInstancePublicPortsRequest

PutInstancePublicPortsResult

RebootInstanceRequest

RebootInstanceResult

RebootRelationalDatabaseRequest

RebootRelationalDatabaseResult

Region
Describes the AWS Region.

RelationalDatabase
Describes a database.

RelationalDatabaseBlueprint
Describes a database image, or blueprint. A blueprint describes the major engine version of a database.

RelationalDatabaseBundle
Describes a database bundle. A bundle describes the performance specifications of the database.

RelationalDatabaseEndpoint
Describes an endpoint for a database.

RelationalDatabaseEvent
Describes an event for a database.

RelationalDatabaseHardware
Describes the hardware of a database.

RelationalDatabaseParameter
Describes the parameters of a database.

RelationalDatabaseSnapshot
Describes a database snapshot.

ReleaseStaticIpRequest

ReleaseStaticIpResult

ResourceLocation
Describes the resource location.

StartInstanceRequest

StartInstanceResult

StartRelationalDatabaseRequest

StartRelationalDatabaseResult

StaticIp
Describes the static IP.

StopInstanceRequest

StopInstanceResult

StopRelationalDatabaseRequest

StopRelationalDatabaseResult

Tag
Describes a tag key and optional value assigned to an Amazon Lightsail resource.

For more information about tags in Lightsail, see the Lightsail Dev Guide.

TagResourceRequest

TagResourceResult

UnpeerVpcRequest

UnpeerVpcResult

UntagResourceRequest

UntagResourceResult

UpdateDomainEntryRequest

UpdateDomainEntryResult

UpdateLoadBalancerAttributeRequest

UpdateLoadBalancerAttributeResult

UpdateRelationalDatabaseParametersRequest

UpdateRelationalDatabaseParametersResult

UpdateRelationalDatabaseRequest

UpdateRelationalDatabaseResult

Enums

AllocateStaticIpError
Errors returned by AllocateStaticIp

AttachDiskError
Errors returned by AttachDisk

AttachInstancesToLoadBalancerError
Errors returned by AttachInstancesToLoadBalancer

AttachLoadBalancerTlsCertificateError
Errors returned by AttachLoadBalancerTlsCertificate

AttachStaticIpError
Errors returned by AttachStaticIp

CloseInstancePublicPortsError
Errors returned by CloseInstancePublicPorts

CopySnapshotError
Errors returned by CopySnapshot

CreateCloudFormationStackError
Errors returned by CreateCloudFormationStack

CreateDiskError
Errors returned by CreateDisk

CreateDiskFromSnapshotError
Errors returned by CreateDiskFromSnapshot

CreateDiskSnapshotError
Errors returned by CreateDiskSnapshot

CreateDomainEntryError
Errors returned by CreateDomainEntry

CreateDomainError
Errors returned by CreateDomain

CreateInstanceSnapshotError
Errors returned by CreateInstanceSnapshot

CreateInstancesError
Errors returned by CreateInstances

CreateInstancesFromSnapshotError
Errors returned by CreateInstancesFromSnapshot

CreateKeyPairError
Errors returned by CreateKeyPair

CreateLoadBalancerError
Errors returned by CreateLoadBalancer

CreateLoadBalancerTlsCertificateError
Errors returned by CreateLoadBalancerTlsCertificate

CreateRelationalDatabaseError
Errors returned by CreateRelationalDatabase

CreateRelationalDatabaseFromSnapshotError
Errors returned by CreateRelationalDatabaseFromSnapshot

CreateRelationalDatabaseSnapshotError
Errors returned by CreateRelationalDatabaseSnapshot

DeleteDiskError
Errors returned by DeleteDisk

DeleteDiskSnapshotError
Errors returned by DeleteDiskSnapshot

DeleteDomainEntryError
Errors returned by DeleteDomainEntry

DeleteDomainError
Errors returned by DeleteDomain

DeleteInstanceError
Errors returned by DeleteInstance

DeleteInstanceSnapshotError
Errors returned by DeleteInstanceSnapshot

DeleteKeyPairError
Errors returned by DeleteKeyPair

DeleteKnownHostKeysError
Errors returned by DeleteKnownHostKeys

DeleteLoadBalancerError
Errors returned by DeleteLoadBalancer

DeleteLoadBalancerTlsCertificateError
Errors returned by DeleteLoadBalancerTlsCertificate

DeleteRelationalDatabaseError
Errors returned by DeleteRelationalDatabase

DeleteRelationalDatabaseSnapshotError
Errors returned by DeleteRelationalDatabaseSnapshot

DetachDiskError
Errors returned by DetachDisk

DetachInstancesFromLoadBalancerError
Errors returned by DetachInstancesFromLoadBalancer

DetachStaticIpError
Errors returned by DetachStaticIp

DownloadDefaultKeyPairError
Errors returned by DownloadDefaultKeyPair

ExportSnapshotError
Errors returned by ExportSnapshot

GetActiveNamesError
Errors returned by GetActiveNames

GetBlueprintsError
Errors returned by GetBlueprints

GetBundlesError
Errors returned by GetBundles

GetCloudFormationStackRecordsError
Errors returned by GetCloudFormationStackRecords

GetDiskError
Errors returned by GetDisk

GetDiskSnapshotError
Errors returned by GetDiskSnapshot

GetDiskSnapshotsError
Errors returned by GetDiskSnapshots

GetDisksError
Errors returned by GetDisks

GetDomainError
Errors returned by GetDomain

GetDomainsError
Errors returned by GetDomains

GetExportSnapshotRecordsError
Errors returned by GetExportSnapshotRecords

GetInstanceAccessDetailsError
Errors returned by GetInstanceAccessDetails

GetInstanceError
Errors returned by GetInstance

GetInstanceMetricDataError
Errors returned by GetInstanceMetricData

GetInstancePortStatesError
Errors returned by GetInstancePortStates

GetInstanceSnapshotError
Errors returned by GetInstanceSnapshot

GetInstanceSnapshotsError
Errors returned by GetInstanceSnapshots

GetInstanceStateError
Errors returned by GetInstanceState

GetInstancesError
Errors returned by GetInstances

GetKeyPairError
Errors returned by GetKeyPair

GetKeyPairsError
Errors returned by GetKeyPairs

GetLoadBalancerError
Errors returned by GetLoadBalancer

GetLoadBalancerMetricDataError
Errors returned by GetLoadBalancerMetricData

GetLoadBalancerTlsCertificatesError
Errors returned by GetLoadBalancerTlsCertificates

GetLoadBalancersError
Errors returned by GetLoadBalancers

GetOperationError
Errors returned by GetOperation

GetOperationsError
Errors returned by GetOperations

GetOperationsForResourceError
Errors returned by GetOperationsForResource

GetRegionsError
Errors returned by GetRegions

GetRelationalDatabaseBlueprintsError
Errors returned by GetRelationalDatabaseBlueprints

GetRelationalDatabaseBundlesError
Errors returned by GetRelationalDatabaseBundles

GetRelationalDatabaseError
Errors returned by GetRelationalDatabase

GetRelationalDatabaseEventsError
Errors returned by GetRelationalDatabaseEvents

GetRelationalDatabaseLogEventsError
Errors returned by GetRelationalDatabaseLogEvents

GetRelationalDatabaseLogStreamsError
Errors returned by GetRelationalDatabaseLogStreams

GetRelationalDatabaseMasterUserPasswordError
Errors returned by GetRelationalDatabaseMasterUserPassword

GetRelationalDatabaseMetricDataError
Errors returned by GetRelationalDatabaseMetricData

GetRelationalDatabaseParametersError
Errors returned by GetRelationalDatabaseParameters

GetRelationalDatabaseSnapshotError
Errors returned by GetRelationalDatabaseSnapshot

GetRelationalDatabaseSnapshotsError
Errors returned by GetRelationalDatabaseSnapshots

GetRelationalDatabasesError
Errors returned by GetRelationalDatabases

GetStaticIpError
Errors returned by GetStaticIp

GetStaticIpsError
Errors returned by GetStaticIps

ImportKeyPairError
Errors returned by ImportKeyPair

IsVpcPeeredError
Errors returned by IsVpcPeered

OpenInstancePublicPortsError
Errors returned by OpenInstancePublicPorts

PeerVpcError
Errors returned by PeerVpc

PutInstancePublicPortsError
Errors returned by PutInstancePublicPorts

RebootInstanceError
Errors returned by RebootInstance

RebootRelationalDatabaseError
Errors returned by RebootRelationalDatabase

ReleaseStaticIpError
Errors returned by ReleaseStaticIp

StartInstanceError
Errors returned by StartInstance

StartRelationalDatabaseError
Errors returned by StartRelationalDatabase

StopInstanceError
Errors returned by StopInstance

StopRelationalDatabaseError
Errors returned by StopRelationalDatabase

TagResourceError
Errors returned by TagResource

UnpeerVpcError
Errors returned by UnpeerVpc

UntagResourceError
Errors returned by UntagResource

UpdateDomainEntryError
Errors returned by UpdateDomainEntry

UpdateLoadBalancerAttributeError
Errors returned by UpdateLoadBalancerAttribute

UpdateRelationalDatabaseError
Errors returned by UpdateRelationalDatabase

UpdateRelationalDatabaseParametersError
Errors returned by UpdateRelationalDatabaseParameters

Traits

Lightsail
Trait representing the capabilities of the Amazon Lightsail API. Amazon Lightsail clients implement this trait.