Struct rusoto_lightsail::CreateInstancesRequest

pub struct CreateInstancesRequest {
    pub availability_zone: String,
    pub blueprint_id: String,
    pub bundle_id: String,
    pub instance_names: Vec<String>,
    pub key_pair_name: Option<String>,
    pub user_data: Option<String>,
}

Fields
availability_zone: String

The Availability Zone in which to create your instance. Use the following format: us-east-2a (case sensitive). You can get a list of availability zones by using the get regions operation. Be sure to add the include availability zones parameter to your request.

blueprint_id: String

The ID for a virtual private server image (e.g., app_wordpress_4_4 or app_lamp_7_0). Use the get blueprints operation to return a list of available images (or blueprints).

bundle_id: String

The bundle of specification information for your virtual private server (or instance), including the pricing plan (e.g., micro_1_0).

instance_names: Vec<String>

The names to use for your new Lightsail instances. Separate multiple values using quotation marks and commas, for example: ["MyFirstInstance","MySecondInstance"]

key_pair_name: Option<String>

The name of your key pair.

user_data: Option<String>

A launch script you can create that configures a server with additional user data. For example, you might want to run apt-get -y update.

Depending on the machine image you choose, the command to get software on your instance varies. Amazon Linux and CentOS use yum, Debian and Ubuntu use apt-get, and FreeBSD uses pkg. For a complete list, see the Dev Guide.