Struct rusoto_lightsail::Disk

pub struct Disk {
    pub arn: Option<String>,
    pub attached_to: Option<String>,
    pub attachment_state: Option<String>,
    pub created_at: Option<f64>,
    pub gb_in_use: Option<i64>,
    pub iops: Option<i64>,
    pub is_attached: Option<bool>,
    pub is_system_disk: Option<bool>,
    pub location: Option<ResourceLocation>,
    pub name: Option<String>,
    pub path: Option<String>,
    pub resource_type: Option<String>,
    pub size_in_gb: Option<i64>,
    pub support_code: Option<String>,
}

Describes the hard disk (an SSD).

Fields
arn: Option<String>

The Amazon Resource Name (ARN) of the disk.

attached_to: Option<String>

The resources to which the disk is attached.

attachment_state: Option<String>

The attachment state of the disk.

created_at: Option<f64>

The date when the disk was created.

gb_in_use: Option<i64>

The number of GB in use by the disk.

iops: Option<i64>

The input/output operations per second (IOPS) of the disk.

is_attached: Option<bool>

A Boolean value indicating whether the disk is attached.

is_system_disk: Option<bool>

A Boolean value indicating whether this disk is a system disk (has an operating system loaded on it).

location: Option<ResourceLocation>

The region and Availability Zone where the disk is located.

name: Option<String>

The name of the disk.

path: Option<String>

The disk path.

resource_type: Option<String>

The resource type of the disk.

size_in_gb: Option<i64>

The size of the disk in GB.

support_code: Option<String>

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.