Struct rusoto_lightsail::Instance

pub struct Instance {
    pub arn: Option<String>,
    pub blueprint_id: Option<String>,
    pub blueprint_name: Option<String>,
    pub bundle_id: Option<String>,
    pub created_at: Option<f64>,
    pub hardware: Option<InstanceHardware>,
    pub ipv_6_address: Option<String>,
    pub is_static_ip: Option<bool>,
    pub location: Option<ResourceLocation>,
    pub name: Option<String>,
    pub networking: Option<InstanceNetworking>,
    pub private_ip_address: Option<String>,
    pub public_ip_address: Option<String>,
    pub resource_type: Option<String>,
    pub ssh_key_name: Option<String>,
    pub state: Option<InstanceState>,
    pub support_code: Option<String>,
    pub username: Option<String>,
}

Describes an instance (a virtual private server).

Fields
arn: Option<String>

The Amazon Resource Name (ARN) of the instance (e.g., arn:aws:lightsail:us-east-1:123456789101:Instance/244ad76f-8aad-4741-809f-12345EXAMPLE).

blueprint_id: Option<String>

The blueprint ID (e.g., os_amlinux_2016_03).

blueprint_name: Option<String>

The friendly name of the blueprint (e.g., Amazon Linux).

bundle_id: Option<String>

The bundle for the instance (e.g., micro_1_0).

created_at: Option<f64>

The timestamp when the instance was created (e.g., 1479734909.17).

hardware: Option<InstanceHardware>

The size of the vCPU and the amount of RAM for the instance.

ipv_6_address: Option<String>

The IPv6 address of the instance.

is_static_ip: Option<bool>

A Boolean value indicating whether this instance has a static IP assigned to it.

location: Option<ResourceLocation>

The region name and availability zone where the instance is located.

name: Option<String>

The name the user gave the instance (e.g., Amazon_Linux-1GB-Virginia-1).

networking: Option<InstanceNetworking>

Information about the public ports and monthly data transfer rates for the instance.

private_ip_address: Option<String>

The private IP address of the instance.

public_ip_address: Option<String>

The public IP address of the instance.

resource_type: Option<String>

The type of resource (usually Instance).

ssh_key_name: Option<String>

The name of the SSH key being used to connect to the instance (e.g., LightsailDefaultKeyPair).

state: Option<InstanceState>

The status code and the state (e.g., running) for the instance.

support_code: Option<String>

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

username: Option<String>

The user name for connecting to the instance (e.g., ec2-user).