Struct rusoto_lightsail::InstancePortInfo

pub struct InstancePortInfo {
    pub access_direction: Option<String>,
    pub access_from: Option<String>,
    pub access_type: Option<String>,
    pub common_name: Option<String>,
    pub from_port: Option<i64>,
    pub protocol: Option<String>,
    pub to_port: Option<i64>,
}

Describes information about the instance ports.

Fields
access_direction: Option<String>

The access direction (inbound or outbound).

access_from: Option<String>

The location from which access is allowed (e.g., Anywhere (0.0.0.0/0)).

access_type: Option<String>

The type of access (Public or Private).

common_name: Option<String>

The common name.

from_port: Option<i64>

The first port in the range.

protocol: Option<String>

The protocol being used. Can be one of the following.

tcp - Transmission Control Protocol (TCP) provides reliable, ordered, and error-checked delivery of streamed data between applications running on hosts communicating by an IP network. If you have an application that doesn't require reliable data stream service, use UDP instead.

all - All transport layer protocol types. For more general information, see Transport layer on Wikipedia.

udp - With User Datagram Protocol (UDP), computer applications can send messages (or datagrams) to other hosts on an Internet Protocol (IP) network. Prior communications are not required to set up transmission channels or data paths. Applications that don't require reliable data stream service can use UDP, which provides a connectionless datagram service that emphasizes reduced latency over reliability. If you do require reliable data stream service, use TCP instead.

to_port: Option<i64>

The last port in the range.