Struct rusoto_lightsail::LightsailClient

pub struct LightsailClient { /* fields omitted */ }

A client for the Amazon Lightsail API.

Methods

impl LightsailClient

pub fn new(region: Region) -> LightsailClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> LightsailClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,