Struct rusoto_lightsail::LoadBalancer

pub struct LoadBalancer {
    pub arn: Option<String>,
    pub configuration_options: Option<HashMap<String, String>>,
    pub created_at: Option<f64>,
    pub dns_name: Option<String>,
    pub health_check_path: Option<String>,
    pub instance_health_summary: Option<Vec<InstanceHealthSummary>>,
    pub instance_port: Option<i64>,
    pub location: Option<ResourceLocation>,
    pub name: Option<String>,
    pub protocol: Option<String>,
    pub public_ports: Option<Vec<i64>>,
    pub resource_type: Option<String>,
    pub state: Option<String>,
    pub support_code: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub tls_certificate_summaries: Option<Vec<LoadBalancerTlsCertificateSummary>>,
}

Describes the Lightsail load balancer.

Fields
arn: Option<String>

The Amazon Resource Name (ARN) of the load balancer.

configuration_options: Option<HashMap<String, String>>

A string to string map of the configuration options for your load balancer. Valid values are listed below.

created_at: Option<f64>

The date when your load balancer was created.

dns_name: Option<String>

The DNS name of your Lightsail load balancer.

health_check_path: Option<String>

The path you specified to perform your health checks. If no path is specified, the load balancer tries to make a request to the default (root) page.

instance_health_summary: Option<Vec<InstanceHealthSummary>>

An array of InstanceHealthSummary objects describing the health of the load balancer.

instance_port: Option<i64>

The port where the load balancer will direct traffic to your Lightsail instances. For HTTP traffic, it's port 80. For HTTPS traffic, it's port 443.

location: Option<ResourceLocation>

The AWS Region where your load balancer was created (e.g., us-east-2a). Lightsail automatically creates your load balancer across Availability Zones.

name: Option<String>

The name of the load balancer (e.g., my-load-balancer).

protocol: Option<String>

The protocol you have enabled for your load balancer. Valid values are below.

You can't just have HTTP_HTTPS, but you can have just HTTP.

public_ports: Option<Vec<i64>>

An array of public port settings for your load balancer. For HTTP, use port 80. For HTTPS, use port 443.

resource_type: Option<String>

The resource type (e.g., LoadBalancer.

state: Option<String>

The status of your load balancer. Valid values are below.

support_code: Option<String>

The support code. Include this code in your email to support when you have questions about your Lightsail load balancer. This code enables our support team to look up your Lightsail information more easily.

tags: Option<Vec<Tag>>

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Lightsail Dev Guide.

tls_certificate_summaries: Option<Vec<LoadBalancerTlsCertificateSummary>>

An array of LoadBalancerTlsCertificateSummary objects that provide additional information about the SSL/TLS certificates. For example, if true, the certificate is attached to the load balancer.