Trait rusoto_lightsail::Lightsail

pub trait Lightsail {
    fn allocate_static_ip(
        &self, 
        input: AllocateStaticIpRequest
    ) -> RusotoFuture<AllocateStaticIpResult, AllocateStaticIpError>;

    fn attach_disk(
        &self, 
        input: AttachDiskRequest
    ) -> RusotoFuture<AttachDiskResult, AttachDiskError>;

    fn attach_instances_to_load_balancer(
        &self, 
        input: AttachInstancesToLoadBalancerRequest
    ) -> RusotoFuture<AttachInstancesToLoadBalancerResult, AttachInstancesToLoadBalancerError>;

    fn attach_load_balancer_tls_certificate(
        &self, 
        input: AttachLoadBalancerTlsCertificateRequest
    ) -> RusotoFuture<AttachLoadBalancerTlsCertificateResult, AttachLoadBalancerTlsCertificateError>;

    fn attach_static_ip(
        &self, 
        input: AttachStaticIpRequest
    ) -> RusotoFuture<AttachStaticIpResult, AttachStaticIpError>;

    fn close_instance_public_ports(
        &self, 
        input: CloseInstancePublicPortsRequest
    ) -> RusotoFuture<CloseInstancePublicPortsResult, CloseInstancePublicPortsError>;

    fn copy_snapshot(
        &self, 
        input: CopySnapshotRequest
    ) -> RusotoFuture<CopySnapshotResult, CopySnapshotError>;

    fn create_cloud_formation_stack(
        &self, 
        input: CreateCloudFormationStackRequest
    ) -> RusotoFuture<CreateCloudFormationStackResult, CreateCloudFormationStackError>;

    fn create_disk(
        &self, 
        input: CreateDiskRequest
    ) -> RusotoFuture<CreateDiskResult, CreateDiskError>;

    fn create_disk_from_snapshot(
        &self, 
        input: CreateDiskFromSnapshotRequest
    ) -> RusotoFuture<CreateDiskFromSnapshotResult, CreateDiskFromSnapshotError>;

    fn create_disk_snapshot(
        &self, 
        input: CreateDiskSnapshotRequest
    ) -> RusotoFuture<CreateDiskSnapshotResult, CreateDiskSnapshotError>;

    fn create_domain(
        &self, 
        input: CreateDomainRequest
    ) -> RusotoFuture<CreateDomainResult, CreateDomainError>;

    fn create_domain_entry(
        &self, 
        input: CreateDomainEntryRequest
    ) -> RusotoFuture<CreateDomainEntryResult, CreateDomainEntryError>;

    fn create_instance_snapshot(
        &self, 
        input: CreateInstanceSnapshotRequest
    ) -> RusotoFuture<CreateInstanceSnapshotResult, CreateInstanceSnapshotError>;

    fn create_instances(
        &self, 
        input: CreateInstancesRequest
    ) -> RusotoFuture<CreateInstancesResult, CreateInstancesError>;

    fn create_instances_from_snapshot(
        &self, 
        input: CreateInstancesFromSnapshotRequest
    ) -> RusotoFuture<CreateInstancesFromSnapshotResult, CreateInstancesFromSnapshotError>;

    fn create_key_pair(
        &self, 
        input: CreateKeyPairRequest
    ) -> RusotoFuture<CreateKeyPairResult, CreateKeyPairError>;

    fn create_load_balancer(
        &self, 
        input: CreateLoadBalancerRequest
    ) -> RusotoFuture<CreateLoadBalancerResult, CreateLoadBalancerError>;

    fn create_load_balancer_tls_certificate(
        &self, 
        input: CreateLoadBalancerTlsCertificateRequest
    ) -> RusotoFuture<CreateLoadBalancerTlsCertificateResult, CreateLoadBalancerTlsCertificateError>;

    fn create_relational_database(
        &self, 
        input: CreateRelationalDatabaseRequest
    ) -> RusotoFuture<CreateRelationalDatabaseResult, CreateRelationalDatabaseError>;

    fn create_relational_database_from_snapshot(
        &self, 
        input: CreateRelationalDatabaseFromSnapshotRequest
    ) -> RusotoFuture<CreateRelationalDatabaseFromSnapshotResult, CreateRelationalDatabaseFromSnapshotError>;

    fn create_relational_database_snapshot(
        &self, 
        input: CreateRelationalDatabaseSnapshotRequest
    ) -> RusotoFuture<CreateRelationalDatabaseSnapshotResult, CreateRelationalDatabaseSnapshotError>;

    fn delete_disk(
        &self, 
        input: DeleteDiskRequest
    ) -> RusotoFuture<DeleteDiskResult, DeleteDiskError>;

    fn delete_disk_snapshot(
        &self, 
        input: DeleteDiskSnapshotRequest
    ) -> RusotoFuture<DeleteDiskSnapshotResult, DeleteDiskSnapshotError>;

    fn delete_domain(
        &self, 
        input: DeleteDomainRequest
    ) -> RusotoFuture<DeleteDomainResult, DeleteDomainError>;

    fn delete_domain_entry(
        &self, 
        input: DeleteDomainEntryRequest
    ) -> RusotoFuture<DeleteDomainEntryResult, DeleteDomainEntryError>;

    fn delete_instance(
        &self, 
        input: DeleteInstanceRequest
    ) -> RusotoFuture<DeleteInstanceResult, DeleteInstanceError>;

    fn delete_instance_snapshot(
        &self, 
        input: DeleteInstanceSnapshotRequest
    ) -> RusotoFuture<DeleteInstanceSnapshotResult, DeleteInstanceSnapshotError>;

    fn delete_key_pair(
        &self, 
        input: DeleteKeyPairRequest
    ) -> RusotoFuture<DeleteKeyPairResult, DeleteKeyPairError>;

    fn delete_known_host_keys(
        &self, 
        input: DeleteKnownHostKeysRequest
    ) -> RusotoFuture<DeleteKnownHostKeysResult, DeleteKnownHostKeysError>;

    fn delete_load_balancer(
        &self, 
        input: DeleteLoadBalancerRequest
    ) -> RusotoFuture<DeleteLoadBalancerResult, DeleteLoadBalancerError>;

    fn delete_load_balancer_tls_certificate(
        &self, 
        input: DeleteLoadBalancerTlsCertificateRequest
    ) -> RusotoFuture<DeleteLoadBalancerTlsCertificateResult, DeleteLoadBalancerTlsCertificateError>;

    fn delete_relational_database(
        &self, 
        input: DeleteRelationalDatabaseRequest
    ) -> RusotoFuture<DeleteRelationalDatabaseResult, DeleteRelationalDatabaseError>;

    fn delete_relational_database_snapshot(
        &self, 
        input: DeleteRelationalDatabaseSnapshotRequest
    ) -> RusotoFuture<DeleteRelationalDatabaseSnapshotResult, DeleteRelationalDatabaseSnapshotError>;

    fn detach_disk(
        &self, 
        input: DetachDiskRequest
    ) -> RusotoFuture<DetachDiskResult, DetachDiskError>;

    fn detach_instances_from_load_balancer(
        &self, 
        input: DetachInstancesFromLoadBalancerRequest
    ) -> RusotoFuture<DetachInstancesFromLoadBalancerResult, DetachInstancesFromLoadBalancerError>;

    fn detach_static_ip(
        &self, 
        input: DetachStaticIpRequest
    ) -> RusotoFuture<DetachStaticIpResult, DetachStaticIpError>;

    fn download_default_key_pair(
        &self
    ) -> RusotoFuture<DownloadDefaultKeyPairResult, DownloadDefaultKeyPairError>;

    fn export_snapshot(
        &self, 
        input: ExportSnapshotRequest
    ) -> RusotoFuture<ExportSnapshotResult, ExportSnapshotError>;

    fn get_active_names(
        &self, 
        input: GetActiveNamesRequest
    ) -> RusotoFuture<GetActiveNamesResult, GetActiveNamesError>;

    fn get_blueprints(
        &self, 
        input: GetBlueprintsRequest
    ) -> RusotoFuture<GetBlueprintsResult, GetBlueprintsError>;

    fn get_bundles(
        &self, 
        input: GetBundlesRequest
    ) -> RusotoFuture<GetBundlesResult, GetBundlesError>;

    fn get_cloud_formation_stack_records(
        &self, 
        input: GetCloudFormationStackRecordsRequest
    ) -> RusotoFuture<GetCloudFormationStackRecordsResult, GetCloudFormationStackRecordsError>;

    fn get_disk(
        &self, 
        input: GetDiskRequest
    ) -> RusotoFuture<GetDiskResult, GetDiskError>;

    fn get_disk_snapshot(
        &self, 
        input: GetDiskSnapshotRequest
    ) -> RusotoFuture<GetDiskSnapshotResult, GetDiskSnapshotError>;

    fn get_disk_snapshots(
        &self, 
        input: GetDiskSnapshotsRequest
    ) -> RusotoFuture<GetDiskSnapshotsResult, GetDiskSnapshotsError>;

    fn get_disks(
        &self, 
        input: GetDisksRequest
    ) -> RusotoFuture<GetDisksResult, GetDisksError>;

    fn get_domain(
        &self, 
        input: GetDomainRequest
    ) -> RusotoFuture<GetDomainResult, GetDomainError>;

    fn get_domains(
        &self, 
        input: GetDomainsRequest
    ) -> RusotoFuture<GetDomainsResult, GetDomainsError>;

    fn get_export_snapshot_records(
        &self, 
        input: GetExportSnapshotRecordsRequest
    ) -> RusotoFuture<GetExportSnapshotRecordsResult, GetExportSnapshotRecordsError>;

    fn get_instance(
        &self, 
        input: GetInstanceRequest
    ) -> RusotoFuture<GetInstanceResult, GetInstanceError>;

    fn get_instance_access_details(
        &self, 
        input: GetInstanceAccessDetailsRequest
    ) -> RusotoFuture<GetInstanceAccessDetailsResult, GetInstanceAccessDetailsError>;

    fn get_instance_metric_data(
        &self, 
        input: GetInstanceMetricDataRequest
    ) -> RusotoFuture<GetInstanceMetricDataResult, GetInstanceMetricDataError>;

    fn get_instance_port_states(
        &self, 
        input: GetInstancePortStatesRequest
    ) -> RusotoFuture<GetInstancePortStatesResult, GetInstancePortStatesError>;

    fn get_instance_snapshot(
        &self, 
        input: GetInstanceSnapshotRequest
    ) -> RusotoFuture<GetInstanceSnapshotResult, GetInstanceSnapshotError>;

    fn get_instance_snapshots(
        &self, 
        input: GetInstanceSnapshotsRequest
    ) -> RusotoFuture<GetInstanceSnapshotsResult, GetInstanceSnapshotsError>;

    fn get_instance_state(
        &self, 
        input: GetInstanceStateRequest
    ) -> RusotoFuture<GetInstanceStateResult, GetInstanceStateError>;

    fn get_instances(
        &self, 
        input: GetInstancesRequest
    ) -> RusotoFuture<GetInstancesResult, GetInstancesError>;

    fn get_key_pair(
        &self, 
        input: GetKeyPairRequest
    ) -> RusotoFuture<GetKeyPairResult, GetKeyPairError>;

    fn get_key_pairs(
        &self, 
        input: GetKeyPairsRequest
    ) -> RusotoFuture<GetKeyPairsResult, GetKeyPairsError>;

    fn get_load_balancer(
        &self, 
        input: GetLoadBalancerRequest
    ) -> RusotoFuture<GetLoadBalancerResult, GetLoadBalancerError>;

    fn get_load_balancer_metric_data(
        &self, 
        input: GetLoadBalancerMetricDataRequest
    ) -> RusotoFuture<GetLoadBalancerMetricDataResult, GetLoadBalancerMetricDataError>;

    fn get_load_balancer_tls_certificates(
        &self, 
        input: GetLoadBalancerTlsCertificatesRequest
    ) -> RusotoFuture<GetLoadBalancerTlsCertificatesResult, GetLoadBalancerTlsCertificatesError>;

    fn get_load_balancers(
        &self, 
        input: GetLoadBalancersRequest
    ) -> RusotoFuture<GetLoadBalancersResult, GetLoadBalancersError>;

    fn get_operation(
        &self, 
        input: GetOperationRequest
    ) -> RusotoFuture<GetOperationResult, GetOperationError>;

    fn get_operations(
        &self, 
        input: GetOperationsRequest
    ) -> RusotoFuture<GetOperationsResult, GetOperationsError>;

    fn get_operations_for_resource(
        &self, 
        input: GetOperationsForResourceRequest
    ) -> RusotoFuture<GetOperationsForResourceResult, GetOperationsForResourceError>;

    fn get_regions(
        &self, 
        input: GetRegionsRequest
    ) -> RusotoFuture<GetRegionsResult, GetRegionsError>;

    fn get_relational_database(
        &self, 
        input: GetRelationalDatabaseRequest
    ) -> RusotoFuture<GetRelationalDatabaseResult, GetRelationalDatabaseError>;

    fn get_relational_database_blueprints(
        &self, 
        input: GetRelationalDatabaseBlueprintsRequest
    ) -> RusotoFuture<GetRelationalDatabaseBlueprintsResult, GetRelationalDatabaseBlueprintsError>;

    fn get_relational_database_bundles(
        &self, 
        input: GetRelationalDatabaseBundlesRequest
    ) -> RusotoFuture<GetRelationalDatabaseBundlesResult, GetRelationalDatabaseBundlesError>;

    fn get_relational_database_events(
        &self, 
        input: GetRelationalDatabaseEventsRequest
    ) -> RusotoFuture<GetRelationalDatabaseEventsResult, GetRelationalDatabaseEventsError>;

    fn get_relational_database_log_events(
        &self, 
        input: GetRelationalDatabaseLogEventsRequest
    ) -> RusotoFuture<GetRelationalDatabaseLogEventsResult, GetRelationalDatabaseLogEventsError>;

    fn get_relational_database_log_streams(
        &self, 
        input: GetRelationalDatabaseLogStreamsRequest
    ) -> RusotoFuture<GetRelationalDatabaseLogStreamsResult, GetRelationalDatabaseLogStreamsError>;

    fn get_relational_database_master_user_password(
        &self, 
        input: GetRelationalDatabaseMasterUserPasswordRequest
    ) -> RusotoFuture<GetRelationalDatabaseMasterUserPasswordResult, GetRelationalDatabaseMasterUserPasswordError>;

    fn get_relational_database_metric_data(
        &self, 
        input: GetRelationalDatabaseMetricDataRequest
    ) -> RusotoFuture<GetRelationalDatabaseMetricDataResult, GetRelationalDatabaseMetricDataError>;

    fn get_relational_database_parameters(
        &self, 
        input: GetRelationalDatabaseParametersRequest
    ) -> RusotoFuture<GetRelationalDatabaseParametersResult, GetRelationalDatabaseParametersError>;

    fn get_relational_database_snapshot(
        &self, 
        input: GetRelationalDatabaseSnapshotRequest
    ) -> RusotoFuture<GetRelationalDatabaseSnapshotResult, GetRelationalDatabaseSnapshotError>;

    fn get_relational_database_snapshots(
        &self, 
        input: GetRelationalDatabaseSnapshotsRequest
    ) -> RusotoFuture<GetRelationalDatabaseSnapshotsResult, GetRelationalDatabaseSnapshotsError>;

    fn get_relational_databases(
        &self, 
        input: GetRelationalDatabasesRequest
    ) -> RusotoFuture<GetRelationalDatabasesResult, GetRelationalDatabasesError>;

    fn get_static_ip(
        &self, 
        input: GetStaticIpRequest
    ) -> RusotoFuture<GetStaticIpResult, GetStaticIpError>;

    fn get_static_ips(
        &self, 
        input: GetStaticIpsRequest
    ) -> RusotoFuture<GetStaticIpsResult, GetStaticIpsError>;

    fn import_key_pair(
        &self, 
        input: ImportKeyPairRequest
    ) -> RusotoFuture<ImportKeyPairResult, ImportKeyPairError>;

    fn is_vpc_peered(&self) -> RusotoFuture<IsVpcPeeredResult, IsVpcPeeredError>;

    fn open_instance_public_ports(
        &self, 
        input: OpenInstancePublicPortsRequest
    ) -> RusotoFuture<OpenInstancePublicPortsResult, OpenInstancePublicPortsError>;

    fn peer_vpc(&self) -> RusotoFuture<PeerVpcResult, PeerVpcError>;

    fn put_instance_public_ports(
        &self, 
        input: PutInstancePublicPortsRequest
    ) -> RusotoFuture<PutInstancePublicPortsResult, PutInstancePublicPortsError>;

    fn reboot_instance(
        &self, 
        input: RebootInstanceRequest
    ) -> RusotoFuture<RebootInstanceResult, RebootInstanceError>;

    fn reboot_relational_database(
        &self, 
        input: RebootRelationalDatabaseRequest
    ) -> RusotoFuture<RebootRelationalDatabaseResult, RebootRelationalDatabaseError>;

    fn release_static_ip(
        &self, 
        input: ReleaseStaticIpRequest
    ) -> RusotoFuture<ReleaseStaticIpResult, ReleaseStaticIpError>;

    fn start_instance(
        &self, 
        input: StartInstanceRequest
    ) -> RusotoFuture<StartInstanceResult, StartInstanceError>;

    fn start_relational_database(
        &self, 
        input: StartRelationalDatabaseRequest
    ) -> RusotoFuture<StartRelationalDatabaseResult, StartRelationalDatabaseError>;

    fn stop_instance(
        &self, 
        input: StopInstanceRequest
    ) -> RusotoFuture<StopInstanceResult, StopInstanceError>;

    fn stop_relational_database(
        &self, 
        input: StopRelationalDatabaseRequest
    ) -> RusotoFuture<StopRelationalDatabaseResult, StopRelationalDatabaseError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<TagResourceResult, TagResourceError>;

    fn unpeer_vpc(&self) -> RusotoFuture<UnpeerVpcResult, UnpeerVpcError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<UntagResourceResult, UntagResourceError>;

    fn update_domain_entry(
        &self, 
        input: UpdateDomainEntryRequest
    ) -> RusotoFuture<UpdateDomainEntryResult, UpdateDomainEntryError>;

    fn update_load_balancer_attribute(
        &self, 
        input: UpdateLoadBalancerAttributeRequest
    ) -> RusotoFuture<UpdateLoadBalancerAttributeResult, UpdateLoadBalancerAttributeError>;

    fn update_relational_database(
        &self, 
        input: UpdateRelationalDatabaseRequest
    ) -> RusotoFuture<UpdateRelationalDatabaseResult, UpdateRelationalDatabaseError>;

    fn update_relational_database_parameters(
        &self, 
        input: UpdateRelationalDatabaseParametersRequest
    ) -> RusotoFuture<UpdateRelationalDatabaseParametersResult, UpdateRelationalDatabaseParametersError>;
}

Trait representing the capabilities of the Amazon Lightsail API. Amazon Lightsail clients implement this trait.

Required methods

fn allocate_static_ip(
&self,
input: AllocateStaticIpRequest
) -> RusotoFuture<AllocateStaticIpResult, AllocateStaticIpError>

Allocates a static IP address.

fn attach_disk(
&self,
input: AttachDiskRequest
) -> RusotoFuture<AttachDiskResult, AttachDiskError>

Attaches a block storage disk to a running or stopped Lightsail instance and exposes it to the instance with the specified disk name.

The attach disk operation supports tag-based access control via resource tags applied to the resource identified by diskName. For more information, see the Lightsail Dev Guide.

fn attach_instances_to_load_balancer(
&self,
input: AttachInstancesToLoadBalancerRequest
) -> RusotoFuture<AttachInstancesToLoadBalancerResult, AttachInstancesToLoadBalancerError>

Attaches one or more Lightsail instances to a load balancer.

After some time, the instances are attached to the load balancer and the health check status is available.

The attach instances to load balancer operation supports tag-based access control via resource tags applied to the resource identified by loadBalancerName. For more information, see the Lightsail Dev Guide.

fn attach_load_balancer_tls_certificate(
&self,
input: AttachLoadBalancerTlsCertificateRequest
) -> RusotoFuture<AttachLoadBalancerTlsCertificateResult, AttachLoadBalancerTlsCertificateError>

Attaches a Transport Layer Security (TLS) certificate to your load balancer. TLS is just an updated, more secure version of Secure Socket Layer (SSL).

Once you create and validate your certificate, you can attach it to your load balancer. You can also use this API to rotate the certificates on your account. Use the AttachLoadBalancerTlsCertificate operation with the non-attached certificate, and it will replace the existing one and become the attached certificate.

The attach load balancer tls certificate operation supports tag-based access control via resource tags applied to the resource identified by loadBalancerName. For more information, see the Lightsail Dev Guide.

fn attach_static_ip(
&self,
input: AttachStaticIpRequest
) -> RusotoFuture<AttachStaticIpResult, AttachStaticIpError>

Attaches a static IP address to a specific Amazon Lightsail instance.

fn close_instance_public_ports(
&self,
input: CloseInstancePublicPortsRequest
) -> RusotoFuture<CloseInstancePublicPortsResult, CloseInstancePublicPortsError>

Closes the public ports on a specific Amazon Lightsail instance.

The close instance public ports operation supports tag-based access control via resource tags applied to the resource identified by instanceName. For more information, see the Lightsail Dev Guide.

fn copy_snapshot(
&self,
input: CopySnapshotRequest
) -> RusotoFuture<CopySnapshotResult, CopySnapshotError>

Copies an instance or disk snapshot from one AWS Region to another in Amazon Lightsail.

fn create_cloud_formation_stack(
&self,
input: CreateCloudFormationStackRequest
) -> RusotoFuture<CreateCloudFormationStackResult, CreateCloudFormationStackError>

Creates an AWS CloudFormation stack, which creates a new Amazon EC2 instance from an exported Amazon Lightsail snapshot. This operation results in a CloudFormation stack record that can be used to track the AWS CloudFormation stack created. Use the get cloud formation stack records operation to get a list of the CloudFormation stacks created.

Wait until after your new Amazon EC2 instance is created before running the create cloud formation stack operation again with the same export snapshot record.

fn create_disk(
&self,
input: CreateDiskRequest
) -> RusotoFuture<CreateDiskResult, CreateDiskError>

Creates a block storage disk that can be attached to a Lightsail instance in the same Availability Zone (e.g., us-east-2a). The disk is created in the regional endpoint that you send the HTTP request to. For more information, see Regions and Availability Zones in Lightsail.

The create disk operation supports tag-based access control via request tags. For more information, see the Lightsail Dev Guide.

fn create_disk_from_snapshot(
&self,
input: CreateDiskFromSnapshotRequest
) -> RusotoFuture<CreateDiskFromSnapshotResult, CreateDiskFromSnapshotError>

Creates a block storage disk from a disk snapshot that can be attached to a Lightsail instance in the same Availability Zone (e.g., us-east-2a). The disk is created in the regional endpoint that you send the HTTP request to. For more information, see Regions and Availability Zones in Lightsail.

The create disk from snapshot operation supports tag-based access control via request tags and resource tags applied to the resource identified by diskSnapshotName. For more information, see the Lightsail Dev Guide.

fn create_disk_snapshot(
&self,
input: CreateDiskSnapshotRequest
) -> RusotoFuture<CreateDiskSnapshotResult, CreateDiskSnapshotError>

Creates a snapshot of a block storage disk. You can use snapshots for backups, to make copies of disks, and to save data before shutting down a Lightsail instance.

You can take a snapshot of an attached disk that is in use; however, snapshots only capture data that has been written to your disk at the time the snapshot command is issued. This may exclude any data that has been cached by any applications or the operating system. If you can pause any file systems on the disk long enough to take a snapshot, your snapshot should be complete. Nevertheless, if you cannot pause all file writes to the disk, you should unmount the disk from within the Lightsail instance, issue the create disk snapshot command, and then remount the disk to ensure a consistent and complete snapshot. You may remount and use your disk while the snapshot status is pending.

You can also use this operation to create a snapshot of an instance's system volume. You might want to do this, for example, to recover data from the system volume of a botched instance or to create a backup of the system volume like you would for a block storage disk. To create a snapshot of a system volume, just define the instance name parameter when issuing the snapshot command, and a snapshot of the defined instance's system volume will be created. After the snapshot is available, you can create a block storage disk from the snapshot and attach it to a running instance to access the data on the disk.

The create disk snapshot operation supports tag-based access control via request tags. For more information, see the Lightsail Dev Guide.

fn create_domain(
&self,
input: CreateDomainRequest
) -> RusotoFuture<CreateDomainResult, CreateDomainError>

Creates a domain resource for the specified domain (e.g., example.com).

The create domain operation supports tag-based access control via request tags. For more information, see the Lightsail Dev Guide.

fn create_domain_entry(
&self,
input: CreateDomainEntryRequest
) -> RusotoFuture<CreateDomainEntryResult, CreateDomainEntryError>

Creates one of the following entry records associated with the domain: Address (A), canonical name (CNAME), mail exchanger (MX), name server (NS), start of authority (SOA), service locator (SRV), or text (TXT).

The create domain entry operation supports tag-based access control via resource tags applied to the resource identified by domainName. For more information, see the Lightsail Dev Guide.

fn create_instance_snapshot(
&self,
input: CreateInstanceSnapshotRequest
) -> RusotoFuture<CreateInstanceSnapshotResult, CreateInstanceSnapshotError>

Creates a snapshot of a specific virtual private server, or instance. You can use a snapshot to create a new instance that is based on that snapshot.

The create instance snapshot operation supports tag-based access control via request tags. For more information, see the Lightsail Dev Guide.

fn create_instances(
&self,
input: CreateInstancesRequest
) -> RusotoFuture<CreateInstancesResult, CreateInstancesError>

Creates one or more Amazon Lightsail virtual private servers, or instances. Create instances using active blueprints. Inactive blueprints are listed to support customers with existing instances but are not necessarily available for launch of new instances. Blueprints are marked inactive when they become outdated due to operating system updates or new application releases. Use the get blueprints operation to return a list of available blueprints.

The create instances operation supports tag-based access control via request tags. For more information, see the Lightsail Dev Guide.

fn create_instances_from_snapshot(
&self,
input: CreateInstancesFromSnapshotRequest
) -> RusotoFuture<CreateInstancesFromSnapshotResult, CreateInstancesFromSnapshotError>

Uses a specific snapshot as a blueprint for creating one or more new instances that are based on that identical configuration.

The create instances from snapshot operation supports tag-based access control via request tags and resource tags applied to the resource identified by instanceSnapshotName. For more information, see the Lightsail Dev Guide.

fn create_key_pair(
&self,
input: CreateKeyPairRequest
) -> RusotoFuture<CreateKeyPairResult, CreateKeyPairError>

Creates an SSH key pair.

The create key pair operation supports tag-based access control via request tags. For more information, see the Lightsail Dev Guide.

fn create_load_balancer(
&self,
input: CreateLoadBalancerRequest
) -> RusotoFuture<CreateLoadBalancerResult, CreateLoadBalancerError>

Creates a Lightsail load balancer. To learn more about deciding whether to load balance your application, see Configure your Lightsail instances for load balancing. You can create up to 5 load balancers per AWS Region in your account.

When you create a load balancer, you can specify a unique name and port settings. To change additional load balancer settings, use the UpdateLoadBalancerAttribute operation.

The create load balancer operation supports tag-based access control via request tags. For more information, see the Lightsail Dev Guide.

fn create_load_balancer_tls_certificate(
&self,
input: CreateLoadBalancerTlsCertificateRequest
) -> RusotoFuture<CreateLoadBalancerTlsCertificateResult, CreateLoadBalancerTlsCertificateError>

Creates a Lightsail load balancer TLS certificate.

TLS is just an updated, more secure version of Secure Socket Layer (SSL).

The create load balancer tls certificate operation supports tag-based access control via resource tags applied to the resource identified by loadBalancerName. For more information, see the Lightsail Dev Guide.

fn create_relational_database(
&self,
input: CreateRelationalDatabaseRequest
) -> RusotoFuture<CreateRelationalDatabaseResult, CreateRelationalDatabaseError>

Creates a new database in Amazon Lightsail.

The create relational database operation supports tag-based access control via request tags. For more information, see the Lightsail Dev Guide.

fn create_relational_database_from_snapshot(
&self,
input: CreateRelationalDatabaseFromSnapshotRequest
) -> RusotoFuture<CreateRelationalDatabaseFromSnapshotResult, CreateRelationalDatabaseFromSnapshotError>

Creates a new database from an existing database snapshot in Amazon Lightsail.

You can create a new database from a snapshot in if something goes wrong with your original database, or to change it to a different plan, such as a high availability or standard plan.

The create relational database from snapshot operation supports tag-based access control via request tags and resource tags applied to the resource identified by relationalDatabaseSnapshotName. For more information, see the Lightsail Dev Guide.

fn create_relational_database_snapshot(
&self,
input: CreateRelationalDatabaseSnapshotRequest
) -> RusotoFuture<CreateRelationalDatabaseSnapshotResult, CreateRelationalDatabaseSnapshotError>

Creates a snapshot of your database in Amazon Lightsail. You can use snapshots for backups, to make copies of a database, and to save data before deleting a database.

The create relational database snapshot operation supports tag-based access control via request tags. For more information, see the Lightsail Dev Guide.

fn delete_disk(
&self,
input: DeleteDiskRequest
) -> RusotoFuture<DeleteDiskResult, DeleteDiskError>

Deletes the specified block storage disk. The disk must be in the available state (not attached to a Lightsail instance).

The disk may remain in the deleting state for several minutes.

The delete disk operation supports tag-based access control via resource tags applied to the resource identified by diskName. For more information, see the Lightsail Dev Guide.

fn delete_disk_snapshot(
&self,
input: DeleteDiskSnapshotRequest
) -> RusotoFuture<DeleteDiskSnapshotResult, DeleteDiskSnapshotError>

Deletes the specified disk snapshot.

When you make periodic snapshots of a disk, the snapshots are incremental, and only the blocks on the device that have changed since your last snapshot are saved in the new snapshot. When you delete a snapshot, only the data not needed for any other snapshot is removed. So regardless of which prior snapshots have been deleted, all active snapshots will have access to all the information needed to restore the disk.

The delete disk snapshot operation supports tag-based access control via resource tags applied to the resource identified by diskSnapshotName. For more information, see the Lightsail Dev Guide.

fn delete_domain(
&self,
input: DeleteDomainRequest
) -> RusotoFuture<DeleteDomainResult, DeleteDomainError>

Deletes the specified domain recordset and all of its domain records.

The delete domain operation supports tag-based access control via resource tags applied to the resource identified by domainName. For more information, see the Lightsail Dev Guide.

fn delete_domain_entry(
&self,
input: DeleteDomainEntryRequest
) -> RusotoFuture<DeleteDomainEntryResult, DeleteDomainEntryError>

Deletes a specific domain entry.

The delete domain entry operation supports tag-based access control via resource tags applied to the resource identified by domainName. For more information, see the Lightsail Dev Guide.

fn delete_instance(
&self,
input: DeleteInstanceRequest
) -> RusotoFuture<DeleteInstanceResult, DeleteInstanceError>

Deletes a specific Amazon Lightsail virtual private server, or instance.

The delete instance operation supports tag-based access control via resource tags applied to the resource identified by instanceName. For more information, see the Lightsail Dev Guide.

fn delete_instance_snapshot(
&self,
input: DeleteInstanceSnapshotRequest
) -> RusotoFuture<DeleteInstanceSnapshotResult, DeleteInstanceSnapshotError>

Deletes a specific snapshot of a virtual private server (or instance).

The delete instance snapshot operation supports tag-based access control via resource tags applied to the resource identified by instanceSnapshotName. For more information, see the Lightsail Dev Guide.

fn delete_key_pair(
&self,
input: DeleteKeyPairRequest
) -> RusotoFuture<DeleteKeyPairResult, DeleteKeyPairError>

Deletes a specific SSH key pair.

The delete key pair operation supports tag-based access control via resource tags applied to the resource identified by keyPairName. For more information, see the Lightsail Dev Guide.

fn delete_known_host_keys(
&self,
input: DeleteKnownHostKeysRequest
) -> RusotoFuture<DeleteKnownHostKeysResult, DeleteKnownHostKeysError>

Deletes the known host key or certificate used by the Amazon Lightsail browser-based SSH or RDP clients to authenticate an instance. This operation enables the Lightsail browser-based SSH or RDP clients to connect to the instance after a host key mismatch.

Perform this operation only if you were expecting the host key or certificate mismatch or if you are familiar with the new host key or certificate on the instance. For more information, see Troubleshooting connection issues when using the Amazon Lightsail browser-based SSH or RDP client.

fn delete_load_balancer(
&self,
input: DeleteLoadBalancerRequest
) -> RusotoFuture<DeleteLoadBalancerResult, DeleteLoadBalancerError>

Deletes a Lightsail load balancer and all its associated SSL/TLS certificates. Once the load balancer is deleted, you will need to create a new load balancer, create a new certificate, and verify domain ownership again.

The delete load balancer operation supports tag-based access control via resource tags applied to the resource identified by loadBalancerName. For more information, see the Lightsail Dev Guide.

fn delete_load_balancer_tls_certificate(
&self,
input: DeleteLoadBalancerTlsCertificateRequest
) -> RusotoFuture<DeleteLoadBalancerTlsCertificateResult, DeleteLoadBalancerTlsCertificateError>

Deletes an SSL/TLS certificate associated with a Lightsail load balancer.

The delete load balancer tls certificate operation supports tag-based access control via resource tags applied to the resource identified by loadBalancerName. For more information, see the Lightsail Dev Guide.

fn delete_relational_database(
&self,
input: DeleteRelationalDatabaseRequest
) -> RusotoFuture<DeleteRelationalDatabaseResult, DeleteRelationalDatabaseError>

Deletes a database in Amazon Lightsail.

The delete relational database operation supports tag-based access control via resource tags applied to the resource identified by relationalDatabaseName. For more information, see the Lightsail Dev Guide.

fn delete_relational_database_snapshot(
&self,
input: DeleteRelationalDatabaseSnapshotRequest
) -> RusotoFuture<DeleteRelationalDatabaseSnapshotResult, DeleteRelationalDatabaseSnapshotError>

Deletes a database snapshot in Amazon Lightsail.

The delete relational database snapshot operation supports tag-based access control via resource tags applied to the resource identified by relationalDatabaseName. For more information, see the Lightsail Dev Guide.

fn detach_disk(
&self,
input: DetachDiskRequest
) -> RusotoFuture<DetachDiskResult, DetachDiskError>

Detaches a stopped block storage disk from a Lightsail instance. Make sure to unmount any file systems on the device within your operating system before stopping the instance and detaching the disk.

The detach disk operation supports tag-based access control via resource tags applied to the resource identified by diskName. For more information, see the Lightsail Dev Guide.

fn detach_instances_from_load_balancer(
&self,
input: DetachInstancesFromLoadBalancerRequest
) -> RusotoFuture<DetachInstancesFromLoadBalancerResult, DetachInstancesFromLoadBalancerError>

Detaches the specified instances from a Lightsail load balancer.

This operation waits until the instances are no longer needed before they are detached from the load balancer.

The detach instances from load balancer operation supports tag-based access control via resource tags applied to the resource identified by loadBalancerName. For more information, see the Lightsail Dev Guide.

fn detach_static_ip(
&self,
input: DetachStaticIpRequest
) -> RusotoFuture<DetachStaticIpResult, DetachStaticIpError>

Detaches a static IP from the Amazon Lightsail instance to which it is attached.

fn download_default_key_pair(
&self
) -> RusotoFuture<DownloadDefaultKeyPairResult, DownloadDefaultKeyPairError>

Downloads the default SSH key pair from the user's account.

fn export_snapshot(
&self,
input: ExportSnapshotRequest
) -> RusotoFuture<ExportSnapshotResult, ExportSnapshotError>

Exports an Amazon Lightsail instance or block storage disk snapshot to Amazon Elastic Compute Cloud (Amazon EC2). This operation results in an export snapshot record that can be used with the create cloud formation stack operation to create new Amazon EC2 instances.

Exported instance snapshots appear in Amazon EC2 as Amazon Machine Images (AMIs), and the instance system disk appears as an Amazon Elastic Block Store (Amazon EBS) volume. Exported disk snapshots appear in Amazon EC2 as Amazon EBS volumes. Snapshots are exported to the same Amazon Web Services Region in Amazon EC2 as the source Lightsail snapshot.

The export snapshot operation supports tag-based access control via resource tags applied to the resource identified by sourceSnapshotName. For more information, see the Lightsail Dev Guide.

Use the get instance snapshots or get disk snapshots operations to get a list of snapshots that you can export to Amazon EC2.

fn get_active_names(
&self,
input: GetActiveNamesRequest
) -> RusotoFuture<GetActiveNamesResult, GetActiveNamesError>

Returns the names of all active (not deleted) resources.

fn get_blueprints(
&self,
input: GetBlueprintsRequest
) -> RusotoFuture<GetBlueprintsResult, GetBlueprintsError>

Returns the list of available instance images, or blueprints. You can use a blueprint to create a new virtual private server already running a specific operating system, as well as a preinstalled app or development stack. The software each instance is running depends on the blueprint image you choose.

fn get_bundles(
&self,
input: GetBundlesRequest
) -> RusotoFuture<GetBundlesResult, GetBundlesError>

Returns the list of bundles that are available for purchase. A bundle describes the specs for your virtual private server (or instance).

fn get_cloud_formation_stack_records(
&self,
input: GetCloudFormationStackRecordsRequest
) -> RusotoFuture<GetCloudFormationStackRecordsResult, GetCloudFormationStackRecordsError>

Returns the CloudFormation stack record created as a result of the create cloud formation stack operation.

An AWS CloudFormation stack is used to create a new Amazon EC2 instance from an exported Lightsail snapshot.

fn get_disk(
&self,
input: GetDiskRequest
) -> RusotoFuture<GetDiskResult, GetDiskError>

Returns information about a specific block storage disk.

fn get_disk_snapshot(
&self,
input: GetDiskSnapshotRequest
) -> RusotoFuture<GetDiskSnapshotResult, GetDiskSnapshotError>

Returns information about a specific block storage disk snapshot.

fn get_disk_snapshots(
&self,
input: GetDiskSnapshotsRequest
) -> RusotoFuture<GetDiskSnapshotsResult, GetDiskSnapshotsError>

Returns information about all block storage disk snapshots in your AWS account and region.

If you are describing a long list of disk snapshots, you can paginate the output to make the list more manageable. You can use the pageToken and nextPageToken values to retrieve the next items in the list.

fn get_disks(
&self,
input: GetDisksRequest
) -> RusotoFuture<GetDisksResult, GetDisksError>

Returns information about all block storage disks in your AWS account and region.

If you are describing a long list of disks, you can paginate the output to make the list more manageable. You can use the pageToken and nextPageToken values to retrieve the next items in the list.

fn get_domain(
&self,
input: GetDomainRequest
) -> RusotoFuture<GetDomainResult, GetDomainError>

Returns information about a specific domain recordset.

fn get_domains(
&self,
input: GetDomainsRequest
) -> RusotoFuture<GetDomainsResult, GetDomainsError>

Returns a list of all domains in the user's account.

fn get_export_snapshot_records(
&self,
input: GetExportSnapshotRecordsRequest
) -> RusotoFuture<GetExportSnapshotRecordsResult, GetExportSnapshotRecordsError>

Returns the export snapshot record created as a result of the export snapshot operation.

An export snapshot record can be used to create a new Amazon EC2 instance and its related resources with the create cloud formation stack operation.

fn get_instance(
&self,
input: GetInstanceRequest
) -> RusotoFuture<GetInstanceResult, GetInstanceError>

Returns information about a specific Amazon Lightsail instance, which is a virtual private server.

fn get_instance_access_details(
&self,
input: GetInstanceAccessDetailsRequest
) -> RusotoFuture<GetInstanceAccessDetailsResult, GetInstanceAccessDetailsError>

Returns temporary SSH keys you can use to connect to a specific virtual private server, or instance.

The get instance access details operation supports tag-based access control via resource tags applied to the resource identified by instanceName. For more information, see the Lightsail Dev Guide.

fn get_instance_metric_data(
&self,
input: GetInstanceMetricDataRequest
) -> RusotoFuture<GetInstanceMetricDataResult, GetInstanceMetricDataError>

Returns the data points for the specified Amazon Lightsail instance metric, given an instance name.

fn get_instance_port_states(
&self,
input: GetInstancePortStatesRequest
) -> RusotoFuture<GetInstancePortStatesResult, GetInstancePortStatesError>

Returns the port states for a specific virtual private server, or instance.

fn get_instance_snapshot(
&self,
input: GetInstanceSnapshotRequest
) -> RusotoFuture<GetInstanceSnapshotResult, GetInstanceSnapshotError>

Returns information about a specific instance snapshot.

fn get_instance_snapshots(
&self,
input: GetInstanceSnapshotsRequest
) -> RusotoFuture<GetInstanceSnapshotsResult, GetInstanceSnapshotsError>

Returns all instance snapshots for the user's account.

fn get_instance_state(
&self,
input: GetInstanceStateRequest
) -> RusotoFuture<GetInstanceStateResult, GetInstanceStateError>

Returns the state of a specific instance. Works on one instance at a time.

fn get_instances(
&self,
input: GetInstancesRequest
) -> RusotoFuture<GetInstancesResult, GetInstancesError>

Returns information about all Amazon Lightsail virtual private servers, or instances.

fn get_key_pair(
&self,
input: GetKeyPairRequest
) -> RusotoFuture<GetKeyPairResult, GetKeyPairError>

Returns information about a specific key pair.

fn get_key_pairs(
&self,
input: GetKeyPairsRequest
) -> RusotoFuture<GetKeyPairsResult, GetKeyPairsError>

Returns information about all key pairs in the user's account.

fn get_load_balancer(
&self,
input: GetLoadBalancerRequest
) -> RusotoFuture<GetLoadBalancerResult, GetLoadBalancerError>

Returns information about the specified Lightsail load balancer.

fn get_load_balancer_metric_data(
&self,
input: GetLoadBalancerMetricDataRequest
) -> RusotoFuture<GetLoadBalancerMetricDataResult, GetLoadBalancerMetricDataError>

Returns information about health metrics for your Lightsail load balancer.

fn get_load_balancer_tls_certificates(
&self,
input: GetLoadBalancerTlsCertificatesRequest
) -> RusotoFuture<GetLoadBalancerTlsCertificatesResult, GetLoadBalancerTlsCertificatesError>

Returns information about the TLS certificates that are associated with the specified Lightsail load balancer.

TLS is just an updated, more secure version of Secure Socket Layer (SSL).

You can have a maximum of 2 certificates associated with a Lightsail load balancer. One is active and the other is inactive.

fn get_load_balancers(
&self,
input: GetLoadBalancersRequest
) -> RusotoFuture<GetLoadBalancersResult, GetLoadBalancersError>

Returns information about all load balancers in an account.

If you are describing a long list of load balancers, you can paginate the output to make the list more manageable. You can use the pageToken and nextPageToken values to retrieve the next items in the list.

fn get_operation(
&self,
input: GetOperationRequest
) -> RusotoFuture<GetOperationResult, GetOperationError>

Returns information about a specific operation. Operations include events such as when you create an instance, allocate a static IP, attach a static IP, and so on.

fn get_operations(
&self,
input: GetOperationsRequest
) -> RusotoFuture<GetOperationsResult, GetOperationsError>

Returns information about all operations.

Results are returned from oldest to newest, up to a maximum of 200. Results can be paged by making each subsequent call to GetOperations use the maximum (last) statusChangedAt value from the previous request.

fn get_operations_for_resource(
&self,
input: GetOperationsForResourceRequest
) -> RusotoFuture<GetOperationsForResourceResult, GetOperationsForResourceError>

Gets operations for a specific resource (e.g., an instance or a static IP).

fn get_regions(
&self,
input: GetRegionsRequest
) -> RusotoFuture<GetRegionsResult, GetRegionsError>

Returns a list of all valid regions for Amazon Lightsail. Use the include availability zones parameter to also return the Availability Zones in a region.

fn get_relational_database(
&self,
input: GetRelationalDatabaseRequest
) -> RusotoFuture<GetRelationalDatabaseResult, GetRelationalDatabaseError>

Returns information about a specific database in Amazon Lightsail.

fn get_relational_database_blueprints(
&self,
input: GetRelationalDatabaseBlueprintsRequest
) -> RusotoFuture<GetRelationalDatabaseBlueprintsResult, GetRelationalDatabaseBlueprintsError>

Returns a list of available database blueprints in Amazon Lightsail. A blueprint describes the major engine version of a database.

You can use a blueprint ID to create a new database that runs a specific database engine.

fn get_relational_database_bundles(
&self,
input: GetRelationalDatabaseBundlesRequest
) -> RusotoFuture<GetRelationalDatabaseBundlesResult, GetRelationalDatabaseBundlesError>

Returns the list of bundles that are available in Amazon Lightsail. A bundle describes the performance specifications for a database.

You can use a bundle ID to create a new database with explicit performance specifications.

fn get_relational_database_events(
&self,
input: GetRelationalDatabaseEventsRequest
) -> RusotoFuture<GetRelationalDatabaseEventsResult, GetRelationalDatabaseEventsError>

Returns a list of events for a specific database in Amazon Lightsail.

fn get_relational_database_log_events(
&self,
input: GetRelationalDatabaseLogEventsRequest
) -> RusotoFuture<GetRelationalDatabaseLogEventsResult, GetRelationalDatabaseLogEventsError>

Returns a list of log events for a database in Amazon Lightsail.

fn get_relational_database_log_streams(
&self,
input: GetRelationalDatabaseLogStreamsRequest
) -> RusotoFuture<GetRelationalDatabaseLogStreamsResult, GetRelationalDatabaseLogStreamsError>

Returns a list of available log streams for a specific database in Amazon Lightsail.

fn get_relational_database_master_user_password(
&self,
input: GetRelationalDatabaseMasterUserPasswordRequest
) -> RusotoFuture<GetRelationalDatabaseMasterUserPasswordResult, GetRelationalDatabaseMasterUserPasswordError>

Returns the current, previous, or pending versions of the master user password for a Lightsail database.

The asdf operation GetRelationalDatabaseMasterUserPassword supports tag-based access control via resource tags applied to the resource identified by relationalDatabaseName.

fn get_relational_database_metric_data(
&self,
input: GetRelationalDatabaseMetricDataRequest
) -> RusotoFuture<GetRelationalDatabaseMetricDataResult, GetRelationalDatabaseMetricDataError>

Returns the data points of the specified metric for a database in Amazon Lightsail.

fn get_relational_database_parameters(
&self,
input: GetRelationalDatabaseParametersRequest
) -> RusotoFuture<GetRelationalDatabaseParametersResult, GetRelationalDatabaseParametersError>

Returns all of the runtime parameters offered by the underlying database software, or engine, for a specific database in Amazon Lightsail.

In addition to the parameter names and values, this operation returns other information about each parameter. This information includes whether changes require a reboot, whether the parameter is modifiable, the allowed values, and the data types.

fn get_relational_database_snapshot(
&self,
input: GetRelationalDatabaseSnapshotRequest
) -> RusotoFuture<GetRelationalDatabaseSnapshotResult, GetRelationalDatabaseSnapshotError>

Returns information about a specific database snapshot in Amazon Lightsail.

fn get_relational_database_snapshots(
&self,
input: GetRelationalDatabaseSnapshotsRequest
) -> RusotoFuture<GetRelationalDatabaseSnapshotsResult, GetRelationalDatabaseSnapshotsError>

Returns information about all of your database snapshots in Amazon Lightsail.

fn get_relational_databases(
&self,
input: GetRelationalDatabasesRequest
) -> RusotoFuture<GetRelationalDatabasesResult, GetRelationalDatabasesError>

Returns information about all of your databases in Amazon Lightsail.

fn get_static_ip(
&self,
input: GetStaticIpRequest
) -> RusotoFuture<GetStaticIpResult, GetStaticIpError>

Returns information about a specific static IP.

fn get_static_ips(
&self,
input: GetStaticIpsRequest
) -> RusotoFuture<GetStaticIpsResult, GetStaticIpsError>

Returns information about all static IPs in the user's account.

fn import_key_pair(
&self,
input: ImportKeyPairRequest
) -> RusotoFuture<ImportKeyPairResult, ImportKeyPairError>

Imports a public SSH key from a specific key pair.

fn is_vpc_peered(&self) -> RusotoFuture<IsVpcPeeredResult, IsVpcPeeredError>

Returns a Boolean value indicating whether your Lightsail VPC is peered.

fn open_instance_public_ports(
&self,
input: OpenInstancePublicPortsRequest
) -> RusotoFuture<OpenInstancePublicPortsResult, OpenInstancePublicPortsError>

Adds public ports to an Amazon Lightsail instance.

The open instance public ports operation supports tag-based access control via resource tags applied to the resource identified by instanceName. For more information, see the Lightsail Dev Guide.

fn peer_vpc(&self) -> RusotoFuture<PeerVpcResult, PeerVpcError>

Tries to peer the Lightsail VPC with the user's default VPC.

fn put_instance_public_ports(
&self,
input: PutInstancePublicPortsRequest
) -> RusotoFuture<PutInstancePublicPortsResult, PutInstancePublicPortsError>

Sets the specified open ports for an Amazon Lightsail instance, and closes all ports for every protocol not included in the current request.

The put instance public ports operation supports tag-based access control via resource tags applied to the resource identified by instanceName. For more information, see the Lightsail Dev Guide.

fn reboot_instance(
&self,
input: RebootInstanceRequest
) -> RusotoFuture<RebootInstanceResult, RebootInstanceError>

Restarts a specific instance.

The reboot instance operation supports tag-based access control via resource tags applied to the resource identified by instanceName. For more information, see the Lightsail Dev Guide.

fn reboot_relational_database(
&self,
input: RebootRelationalDatabaseRequest
) -> RusotoFuture<RebootRelationalDatabaseResult, RebootRelationalDatabaseError>

Restarts a specific database in Amazon Lightsail.

The reboot relational database operation supports tag-based access control via resource tags applied to the resource identified by relationalDatabaseName. For more information, see the Lightsail Dev Guide.

fn release_static_ip(
&self,
input: ReleaseStaticIpRequest
) -> RusotoFuture<ReleaseStaticIpResult, ReleaseStaticIpError>

Deletes a specific static IP from your account.

fn start_instance(
&self,
input: StartInstanceRequest
) -> RusotoFuture<StartInstanceResult, StartInstanceError>

Starts a specific Amazon Lightsail instance from a stopped state. To restart an instance, use the reboot instance operation.

When you start a stopped instance, Lightsail assigns a new public IP address to the instance. To use the same IP address after stopping and starting an instance, create a static IP address and attach it to the instance. For more information, see the Lightsail Dev Guide.

The start instance operation supports tag-based access control via resource tags applied to the resource identified by instanceName. For more information, see the Lightsail Dev Guide.

fn start_relational_database(
&self,
input: StartRelationalDatabaseRequest
) -> RusotoFuture<StartRelationalDatabaseResult, StartRelationalDatabaseError>

Starts a specific database from a stopped state in Amazon Lightsail. To restart a database, use the reboot relational database operation.

The start relational database operation supports tag-based access control via resource tags applied to the resource identified by relationalDatabaseName. For more information, see the Lightsail Dev Guide.

fn stop_instance(
&self,
input: StopInstanceRequest
) -> RusotoFuture<StopInstanceResult, StopInstanceError>

Stops a specific Amazon Lightsail instance that is currently running.

When you start a stopped instance, Lightsail assigns a new public IP address to the instance. To use the same IP address after stopping and starting an instance, create a static IP address and attach it to the instance. For more information, see the Lightsail Dev Guide.

The stop instance operation supports tag-based access control via resource tags applied to the resource identified by instanceName. For more information, see the Lightsail Dev Guide.

fn stop_relational_database(
&self,
input: StopRelationalDatabaseRequest
) -> RusotoFuture<StopRelationalDatabaseResult, StopRelationalDatabaseError>

Stops a specific database that is currently running in Amazon Lightsail.

The stop relational database operation supports tag-based access control via resource tags applied to the resource identified by relationalDatabaseName. For more information, see the Lightsail Dev Guide.

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<TagResourceResult, TagResourceError>

Adds one or more tags to the specified Amazon Lightsail resource. Each resource can have a maximum of 50 tags. Each tag consists of a key and an optional value. Tag keys must be unique per resource. For more information about tags, see the Lightsail Dev Guide.

The tag resource operation supports tag-based access control via request tags and resource tags applied to the resource identified by resourceName. For more information, see the Lightsail Dev Guide.

fn unpeer_vpc(&self) -> RusotoFuture<UnpeerVpcResult, UnpeerVpcError>

Attempts to unpeer the Lightsail VPC from the user's default VPC.

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<UntagResourceResult, UntagResourceError>

Deletes the specified set of tag keys and their values from the specified Amazon Lightsail resource.

The untag resource operation supports tag-based access control via request tags and resource tags applied to the resource identified by resourceName. For more information, see the Lightsail Dev Guide.

fn update_domain_entry(
&self,
input: UpdateDomainEntryRequest
) -> RusotoFuture<UpdateDomainEntryResult, UpdateDomainEntryError>

Updates a domain recordset after it is created.

The update domain entry operation supports tag-based access control via resource tags applied to the resource identified by domainName. For more information, see the Lightsail Dev Guide.

fn update_load_balancer_attribute(
&self,
input: UpdateLoadBalancerAttributeRequest
) -> RusotoFuture<UpdateLoadBalancerAttributeResult, UpdateLoadBalancerAttributeError>

Updates the specified attribute for a load balancer. You can only update one attribute at a time.

The update load balancer attribute operation supports tag-based access control via resource tags applied to the resource identified by loadBalancerName. For more information, see the Lightsail Dev Guide.

fn update_relational_database(
&self,
input: UpdateRelationalDatabaseRequest
) -> RusotoFuture<UpdateRelationalDatabaseResult, UpdateRelationalDatabaseError>

Allows the update of one or more attributes of a database in Amazon Lightsail.

Updates are applied immediately, or in cases where the updates could result in an outage, are applied during the database's predefined maintenance window.

The update relational database operation supports tag-based access control via resource tags applied to the resource identified by relationalDatabaseName. For more information, see the Lightsail Dev Guide.

fn update_relational_database_parameters(
&self,
input: UpdateRelationalDatabaseParametersRequest
) -> RusotoFuture<UpdateRelationalDatabaseParametersResult, UpdateRelationalDatabaseParametersError>

Allows the update of one or more parameters of a database in Amazon Lightsail.

Parameter updates don't cause outages; therefore, their application is not subject to the preferred maintenance window. However, there are two ways in which paramater updates are applied: dynamic or pending-reboot. Parameters marked with a dynamic apply type are applied immediately. Parameters marked with a pending-reboot apply type are applied only after the database is rebooted using the reboot relational database operation.

The update relational database parameters operation supports tag-based access control via resource tags applied to the resource identified by relationalDatabaseName. For more information, see the Lightsail Dev Guide.

Loading content...