Crate rusoto_lightsail

Amazon Lightsail is the easiest way to get started with AWS for developers who just need virtual private servers. Lightsail includes everything you need to launch your project quickly - a virtual machine, SSD-based storage, data transfer, DNS management, and a static IP - for a low, predictable price. You manage those Lightsail servers through the Lightsail console or by using the API or command-line interface (CLI).

For more information about Lightsail concepts and tasks, see the Lightsail Dev Guide.

To use the Lightsail API or the CLI, you will need to use AWS Identity and Access Management (IAM) to generate access keys. For details about how to set this up, see the Lightsail Dev Guide.

If you're using the service, you're probably looking for LightsailClient and Lightsail.

Structs

AllocateStaticIpRequest

AllocateStaticIpResult

AttachDiskRequest

AttachDiskResult

AttachInstancesToLoadBalancerRequest

AttachInstancesToLoadBalancerResult

AttachLoadBalancerTlsCertificateRequest

AttachLoadBalancerTlsCertificateResult

AttachStaticIpRequest

AttachStaticIpResult

AvailabilityZone
Describes an Availability Zone.

Blueprint
Describes a blueprint (a virtual private server image).

Bundle
Describes a bundle, which is a set of specs describing your virtual private server (or instance).

CloseInstancePublicPortsRequest

CloseInstancePublicPortsResult

CreateDiskFromSnapshotRequest

CreateDiskFromSnapshotResult

CreateDiskRequest

CreateDiskResult

CreateDiskSnapshotRequest

CreateDiskSnapshotResult

CreateDomainEntryRequest

CreateDomainEntryResult

CreateDomainRequest

CreateDomainResult

CreateInstanceSnapshotRequest

CreateInstanceSnapshotResult

CreateInstancesFromSnapshotRequest

CreateInstancesFromSnapshotResult

CreateInstancesRequest

CreateInstancesResult

CreateKeyPairRequest

CreateKeyPairResult

CreateLoadBalancerRequest

CreateLoadBalancerResult

CreateLoadBalancerTlsCertificateRequest

CreateLoadBalancerTlsCertificateResult

DeleteDiskRequest

DeleteDiskResult

DeleteDiskSnapshotRequest

DeleteDiskSnapshotResult

DeleteDomainEntryRequest

DeleteDomainEntryResult

DeleteDomainRequest

DeleteDomainResult

DeleteInstanceRequest

DeleteInstanceResult

DeleteInstanceSnapshotRequest

DeleteInstanceSnapshotResult

DeleteKeyPairRequest

DeleteKeyPairResult

DeleteLoadBalancerRequest

DeleteLoadBalancerResult

DeleteLoadBalancerTlsCertificateRequest

DeleteLoadBalancerTlsCertificateResult

DetachDiskRequest

DetachDiskResult

DetachInstancesFromLoadBalancerRequest

DetachInstancesFromLoadBalancerResult

DetachStaticIpRequest

DetachStaticIpResult

Disk
Describes a system disk or an block storage disk.

DiskMap
Describes a block storage disk mapping.

DiskSnapshot
Describes a block storage disk snapshot.

Domain
Describes a domain where you are storing recordsets in Lightsail.

DomainEntry
Describes a domain recordset entry.

DownloadDefaultKeyPairRequest

DownloadDefaultKeyPairResult

GetActiveNamesRequest

GetActiveNamesResult

GetBlueprintsRequest

GetBlueprintsResult

GetBundlesRequest

GetBundlesResult

GetDiskRequest

GetDiskResult

GetDiskSnapshotRequest

GetDiskSnapshotResult

GetDiskSnapshotsRequest

GetDiskSnapshotsResult

GetDisksRequest

GetDisksResult

GetDomainRequest

GetDomainResult

GetDomainsRequest

GetDomainsResult

GetInstanceAccessDetailsRequest

GetInstanceAccessDetailsResult

GetInstanceMetricDataRequest

GetInstanceMetricDataResult

GetInstancePortStatesRequest

GetInstancePortStatesResult

GetInstanceRequest

GetInstanceResult

GetInstanceSnapshotRequest

GetInstanceSnapshotResult

GetInstanceSnapshotsRequest

GetInstanceSnapshotsResult

GetInstanceStateRequest

GetInstanceStateResult

GetInstancesRequest

GetInstancesResult

GetKeyPairRequest

GetKeyPairResult

GetKeyPairsRequest

GetKeyPairsResult

GetLoadBalancerMetricDataRequest

GetLoadBalancerMetricDataResult

GetLoadBalancerRequest

GetLoadBalancerResult

GetLoadBalancerTlsCertificatesRequest

GetLoadBalancerTlsCertificatesResult

GetLoadBalancersRequest

GetLoadBalancersResult

GetOperationRequest

GetOperationResult

GetOperationsForResourceRequest

GetOperationsForResourceResult

GetOperationsRequest

GetOperationsResult

GetRegionsRequest

GetRegionsResult

GetStaticIpRequest

GetStaticIpResult

GetStaticIpsRequest

GetStaticIpsResult

ImportKeyPairRequest

ImportKeyPairResult

Instance
Describes an instance (a virtual private server).

InstanceAccessDetails
The parameters for gaining temporary access to one of your Amazon Lightsail instances.

InstanceHardware
Describes the hardware for the instance.

InstanceHealthSummary
Describes information about the health of the instance.

InstanceNetworking
Describes monthly data transfer rates and port information for an instance.

InstancePortInfo
Describes information about the instance ports.

InstancePortState
Describes the port state.

InstanceSnapshot
Describes the snapshot of the virtual private server, or instance.

InstanceState
Describes the virtual private server (or instance) status.

IsVpcPeeredRequest

IsVpcPeeredResult

KeyPair
Describes the SSH key pair.

LightsailClient
A client for the Amazon Lightsail API.

LoadBalancer
Describes the Lightsail load balancer.

LoadBalancerTlsCertificate
Describes a load balancer SSL/TLS certificate.

TLS is just an updated, more secure version of Secure Socket Layer (SSL).

LoadBalancerTlsCertificateDomainValidationOption
Contains information about the domain names on an SSL/TLS certificate that you will use to validate domain ownership.

LoadBalancerTlsCertificateDomainValidationRecord
Describes the validation record of each domain name in the SSL/TLS certificate.

LoadBalancerTlsCertificateRenewalSummary
Contains information about the status of Lightsail's managed renewal for the certificate.

LoadBalancerTlsCertificateSummary
Provides a summary of SSL/TLS certificate metadata.

MetricDatapoint
Describes the metric data point.

MonthlyTransfer
Describes the monthly data transfer in and out of your virtual private server (or instance).

OpenInstancePublicPortsRequest

OpenInstancePublicPortsResult

Operation
Describes the API operation.

PasswordData
The password data for the Windows Server-based instance, including the ciphertext and the key pair name.

PeerVpcRequest

PeerVpcResult

PortInfo
Describes information about the ports on your virtual private server (or instance).

PutInstancePublicPortsRequest

PutInstancePublicPortsResult

RebootInstanceRequest

RebootInstanceResult

Region
Describes the AWS Region.

ReleaseStaticIpRequest

ReleaseStaticIpResult

ResourceLocation
Describes the resource location.

StartInstanceRequest

StartInstanceResult

StaticIp
Describes the static IP.

StopInstanceRequest

StopInstanceResult

UnpeerVpcRequest

UnpeerVpcResult

UpdateDomainEntryRequest

UpdateDomainEntryResult

UpdateLoadBalancerAttributeRequest

UpdateLoadBalancerAttributeResult

Enums

AllocateStaticIpError
Errors returned by AllocateStaticIp

AttachDiskError
Errors returned by AttachDisk

AttachInstancesToLoadBalancerError
Errors returned by AttachInstancesToLoadBalancer

AttachLoadBalancerTlsCertificateError
Errors returned by AttachLoadBalancerTlsCertificate

AttachStaticIpError
Errors returned by AttachStaticIp

CloseInstancePublicPortsError
Errors returned by CloseInstancePublicPorts

CreateDiskError
Errors returned by CreateDisk

CreateDiskFromSnapshotError
Errors returned by CreateDiskFromSnapshot

CreateDiskSnapshotError
Errors returned by CreateDiskSnapshot

CreateDomainEntryError
Errors returned by CreateDomainEntry

CreateDomainError
Errors returned by CreateDomain

CreateInstanceSnapshotError
Errors returned by CreateInstanceSnapshot

CreateInstancesError
Errors returned by CreateInstances

CreateInstancesFromSnapshotError
Errors returned by CreateInstancesFromSnapshot

CreateKeyPairError
Errors returned by CreateKeyPair

CreateLoadBalancerError
Errors returned by CreateLoadBalancer

CreateLoadBalancerTlsCertificateError
Errors returned by CreateLoadBalancerTlsCertificate

DeleteDiskError
Errors returned by DeleteDisk

DeleteDiskSnapshotError
Errors returned by DeleteDiskSnapshot

DeleteDomainEntryError
Errors returned by DeleteDomainEntry

DeleteDomainError
Errors returned by DeleteDomain

DeleteInstanceError
Errors returned by DeleteInstance

DeleteInstanceSnapshotError
Errors returned by DeleteInstanceSnapshot

DeleteKeyPairError
Errors returned by DeleteKeyPair

DeleteLoadBalancerError
Errors returned by DeleteLoadBalancer

DeleteLoadBalancerTlsCertificateError
Errors returned by DeleteLoadBalancerTlsCertificate

DetachDiskError
Errors returned by DetachDisk

DetachInstancesFromLoadBalancerError
Errors returned by DetachInstancesFromLoadBalancer

DetachStaticIpError
Errors returned by DetachStaticIp

DownloadDefaultKeyPairError
Errors returned by DownloadDefaultKeyPair

GetActiveNamesError
Errors returned by GetActiveNames

GetBlueprintsError
Errors returned by GetBlueprints

GetBundlesError
Errors returned by GetBundles

GetDiskError
Errors returned by GetDisk

GetDiskSnapshotError
Errors returned by GetDiskSnapshot

GetDiskSnapshotsError
Errors returned by GetDiskSnapshots

GetDisksError
Errors returned by GetDisks

GetDomainError
Errors returned by GetDomain

GetDomainsError
Errors returned by GetDomains

GetInstanceAccessDetailsError
Errors returned by GetInstanceAccessDetails

GetInstanceError
Errors returned by GetInstance

GetInstanceMetricDataError
Errors returned by GetInstanceMetricData

GetInstancePortStatesError
Errors returned by GetInstancePortStates

GetInstanceSnapshotError
Errors returned by GetInstanceSnapshot

GetInstanceSnapshotsError
Errors returned by GetInstanceSnapshots

GetInstanceStateError
Errors returned by GetInstanceState

GetInstancesError
Errors returned by GetInstances

GetKeyPairError
Errors returned by GetKeyPair

GetKeyPairsError
Errors returned by GetKeyPairs

GetLoadBalancerError
Errors returned by GetLoadBalancer

GetLoadBalancerMetricDataError
Errors returned by GetLoadBalancerMetricData

GetLoadBalancerTlsCertificatesError
Errors returned by GetLoadBalancerTlsCertificates

GetLoadBalancersError
Errors returned by GetLoadBalancers

GetOperationError
Errors returned by GetOperation

GetOperationsError
Errors returned by GetOperations

GetOperationsForResourceError
Errors returned by GetOperationsForResource

GetRegionsError
Errors returned by GetRegions

GetStaticIpError
Errors returned by GetStaticIp

GetStaticIpsError
Errors returned by GetStaticIps

ImportKeyPairError
Errors returned by ImportKeyPair

IsVpcPeeredError
Errors returned by IsVpcPeered

OpenInstancePublicPortsError
Errors returned by OpenInstancePublicPorts

PeerVpcError
Errors returned by PeerVpc

PutInstancePublicPortsError
Errors returned by PutInstancePublicPorts

RebootInstanceError
Errors returned by RebootInstance

ReleaseStaticIpError
Errors returned by ReleaseStaticIp

StartInstanceError
Errors returned by StartInstance

StopInstanceError
Errors returned by StopInstance

UnpeerVpcError
Errors returned by UnpeerVpc

UpdateDomainEntryError
Errors returned by UpdateDomainEntry

UpdateLoadBalancerAttributeError
Errors returned by UpdateLoadBalancerAttribute

Traits

Lightsail
Trait representing the capabilities of the Amazon Lightsail API. Amazon Lightsail clients implement this trait.