Trait rusoto_lightsail::Lightsail

pub trait Lightsail {
    fn allocate_static_ip(
        &self, 
        input: AllocateStaticIpRequest
    ) -> RusotoFuture<AllocateStaticIpResult, AllocateStaticIpError>;

    fn attach_disk(
        &self, 
        input: AttachDiskRequest
    ) -> RusotoFuture<AttachDiskResult, AttachDiskError>;

    fn attach_instances_to_load_balancer(
        &self, 
        input: AttachInstancesToLoadBalancerRequest
    ) -> RusotoFuture<AttachInstancesToLoadBalancerResult, AttachInstancesToLoadBalancerError>;

    fn attach_load_balancer_tls_certificate(
        &self, 
        input: AttachLoadBalancerTlsCertificateRequest
    ) -> RusotoFuture<AttachLoadBalancerTlsCertificateResult, AttachLoadBalancerTlsCertificateError>;

    fn attach_static_ip(
        &self, 
        input: AttachStaticIpRequest
    ) -> RusotoFuture<AttachStaticIpResult, AttachStaticIpError>;

    fn close_instance_public_ports(
        &self, 
        input: CloseInstancePublicPortsRequest
    ) -> RusotoFuture<CloseInstancePublicPortsResult, CloseInstancePublicPortsError>;

    fn create_disk(
        &self, 
        input: CreateDiskRequest
    ) -> RusotoFuture<CreateDiskResult, CreateDiskError>;

    fn create_disk_from_snapshot(
        &self, 
        input: CreateDiskFromSnapshotRequest
    ) -> RusotoFuture<CreateDiskFromSnapshotResult, CreateDiskFromSnapshotError>;

    fn create_disk_snapshot(
        &self, 
        input: CreateDiskSnapshotRequest
    ) -> RusotoFuture<CreateDiskSnapshotResult, CreateDiskSnapshotError>;

    fn create_domain(
        &self, 
        input: CreateDomainRequest
    ) -> RusotoFuture<CreateDomainResult, CreateDomainError>;

    fn create_domain_entry(
        &self, 
        input: CreateDomainEntryRequest
    ) -> RusotoFuture<CreateDomainEntryResult, CreateDomainEntryError>;

    fn create_instance_snapshot(
        &self, 
        input: CreateInstanceSnapshotRequest
    ) -> RusotoFuture<CreateInstanceSnapshotResult, CreateInstanceSnapshotError>;

    fn create_instances(
        &self, 
        input: CreateInstancesRequest
    ) -> RusotoFuture<CreateInstancesResult, CreateInstancesError>;

    fn create_instances_from_snapshot(
        &self, 
        input: CreateInstancesFromSnapshotRequest
    ) -> RusotoFuture<CreateInstancesFromSnapshotResult, CreateInstancesFromSnapshotError>;

    fn create_key_pair(
        &self, 
        input: CreateKeyPairRequest
    ) -> RusotoFuture<CreateKeyPairResult, CreateKeyPairError>;

    fn create_load_balancer(
        &self, 
        input: CreateLoadBalancerRequest
    ) -> RusotoFuture<CreateLoadBalancerResult, CreateLoadBalancerError>;

    fn create_load_balancer_tls_certificate(
        &self, 
        input: CreateLoadBalancerTlsCertificateRequest
    ) -> RusotoFuture<CreateLoadBalancerTlsCertificateResult, CreateLoadBalancerTlsCertificateError>;

    fn delete_disk(
        &self, 
        input: DeleteDiskRequest
    ) -> RusotoFuture<DeleteDiskResult, DeleteDiskError>;

    fn delete_disk_snapshot(
        &self, 
        input: DeleteDiskSnapshotRequest
    ) -> RusotoFuture<DeleteDiskSnapshotResult, DeleteDiskSnapshotError>;

    fn delete_domain(
        &self, 
        input: DeleteDomainRequest
    ) -> RusotoFuture<DeleteDomainResult, DeleteDomainError>;

    fn delete_domain_entry(
        &self, 
        input: DeleteDomainEntryRequest
    ) -> RusotoFuture<DeleteDomainEntryResult, DeleteDomainEntryError>;

    fn delete_instance(
        &self, 
        input: DeleteInstanceRequest
    ) -> RusotoFuture<DeleteInstanceResult, DeleteInstanceError>;

    fn delete_instance_snapshot(
        &self, 
        input: DeleteInstanceSnapshotRequest
    ) -> RusotoFuture<DeleteInstanceSnapshotResult, DeleteInstanceSnapshotError>;

    fn delete_key_pair(
        &self, 
        input: DeleteKeyPairRequest
    ) -> RusotoFuture<DeleteKeyPairResult, DeleteKeyPairError>;

    fn delete_load_balancer(
        &self, 
        input: DeleteLoadBalancerRequest
    ) -> RusotoFuture<DeleteLoadBalancerResult, DeleteLoadBalancerError>;

    fn delete_load_balancer_tls_certificate(
        &self, 
        input: DeleteLoadBalancerTlsCertificateRequest
    ) -> RusotoFuture<DeleteLoadBalancerTlsCertificateResult, DeleteLoadBalancerTlsCertificateError>;

    fn detach_disk(
        &self, 
        input: DetachDiskRequest
    ) -> RusotoFuture<DetachDiskResult, DetachDiskError>;

    fn detach_instances_from_load_balancer(
        &self, 
        input: DetachInstancesFromLoadBalancerRequest
    ) -> RusotoFuture<DetachInstancesFromLoadBalancerResult, DetachInstancesFromLoadBalancerError>;

    fn detach_static_ip(
        &self, 
        input: DetachStaticIpRequest
    ) -> RusotoFuture<DetachStaticIpResult, DetachStaticIpError>;

    fn download_default_key_pair(
        &self
    ) -> RusotoFuture<DownloadDefaultKeyPairResult, DownloadDefaultKeyPairError>;

    fn get_active_names(
        &self, 
        input: GetActiveNamesRequest
    ) -> RusotoFuture<GetActiveNamesResult, GetActiveNamesError>;

    fn get_blueprints(
        &self, 
        input: GetBlueprintsRequest
    ) -> RusotoFuture<GetBlueprintsResult, GetBlueprintsError>;

    fn get_bundles(
        &self, 
        input: GetBundlesRequest
    ) -> RusotoFuture<GetBundlesResult, GetBundlesError>;

    fn get_disk(
        &self, 
        input: GetDiskRequest
    ) -> RusotoFuture<GetDiskResult, GetDiskError>;

    fn get_disk_snapshot(
        &self, 
        input: GetDiskSnapshotRequest
    ) -> RusotoFuture<GetDiskSnapshotResult, GetDiskSnapshotError>;

    fn get_disk_snapshots(
        &self, 
        input: GetDiskSnapshotsRequest
    ) -> RusotoFuture<GetDiskSnapshotsResult, GetDiskSnapshotsError>;

    fn get_disks(
        &self, 
        input: GetDisksRequest
    ) -> RusotoFuture<GetDisksResult, GetDisksError>;

    fn get_domain(
        &self, 
        input: GetDomainRequest
    ) -> RusotoFuture<GetDomainResult, GetDomainError>;

    fn get_domains(
        &self, 
        input: GetDomainsRequest
    ) -> RusotoFuture<GetDomainsResult, GetDomainsError>;

    fn get_instance(
        &self, 
        input: GetInstanceRequest
    ) -> RusotoFuture<GetInstanceResult, GetInstanceError>;

    fn get_instance_access_details(
        &self, 
        input: GetInstanceAccessDetailsRequest
    ) -> RusotoFuture<GetInstanceAccessDetailsResult, GetInstanceAccessDetailsError>;

    fn get_instance_metric_data(
        &self, 
        input: GetInstanceMetricDataRequest
    ) -> RusotoFuture<GetInstanceMetricDataResult, GetInstanceMetricDataError>;

    fn get_instance_port_states(
        &self, 
        input: GetInstancePortStatesRequest
    ) -> RusotoFuture<GetInstancePortStatesResult, GetInstancePortStatesError>;

    fn get_instance_snapshot(
        &self, 
        input: GetInstanceSnapshotRequest
    ) -> RusotoFuture<GetInstanceSnapshotResult, GetInstanceSnapshotError>;

    fn get_instance_snapshots(
        &self, 
        input: GetInstanceSnapshotsRequest
    ) -> RusotoFuture<GetInstanceSnapshotsResult, GetInstanceSnapshotsError>;

    fn get_instance_state(
        &self, 
        input: GetInstanceStateRequest
    ) -> RusotoFuture<GetInstanceStateResult, GetInstanceStateError>;

    fn get_instances(
        &self, 
        input: GetInstancesRequest
    ) -> RusotoFuture<GetInstancesResult, GetInstancesError>;

    fn get_key_pair(
        &self, 
        input: GetKeyPairRequest
    ) -> RusotoFuture<GetKeyPairResult, GetKeyPairError>;

    fn get_key_pairs(
        &self, 
        input: GetKeyPairsRequest
    ) -> RusotoFuture<GetKeyPairsResult, GetKeyPairsError>;

    fn get_load_balancer(
        &self, 
        input: GetLoadBalancerRequest
    ) -> RusotoFuture<GetLoadBalancerResult, GetLoadBalancerError>;

    fn get_load_balancer_metric_data(
        &self, 
        input: GetLoadBalancerMetricDataRequest
    ) -> RusotoFuture<GetLoadBalancerMetricDataResult, GetLoadBalancerMetricDataError>;

    fn get_load_balancer_tls_certificates(
        &self, 
        input: GetLoadBalancerTlsCertificatesRequest
    ) -> RusotoFuture<GetLoadBalancerTlsCertificatesResult, GetLoadBalancerTlsCertificatesError>;

    fn get_load_balancers(
        &self, 
        input: GetLoadBalancersRequest
    ) -> RusotoFuture<GetLoadBalancersResult, GetLoadBalancersError>;

    fn get_operation(
        &self, 
        input: GetOperationRequest
    ) -> RusotoFuture<GetOperationResult, GetOperationError>;

    fn get_operations(
        &self, 
        input: GetOperationsRequest
    ) -> RusotoFuture<GetOperationsResult, GetOperationsError>;

    fn get_operations_for_resource(
        &self, 
        input: GetOperationsForResourceRequest
    ) -> RusotoFuture<GetOperationsForResourceResult, GetOperationsForResourceError>;

    fn get_regions(
        &self, 
        input: GetRegionsRequest
    ) -> RusotoFuture<GetRegionsResult, GetRegionsError>;

    fn get_static_ip(
        &self, 
        input: GetStaticIpRequest
    ) -> RusotoFuture<GetStaticIpResult, GetStaticIpError>;

    fn get_static_ips(
        &self, 
        input: GetStaticIpsRequest
    ) -> RusotoFuture<GetStaticIpsResult, GetStaticIpsError>;

    fn import_key_pair(
        &self, 
        input: ImportKeyPairRequest
    ) -> RusotoFuture<ImportKeyPairResult, ImportKeyPairError>;

    fn is_vpc_peered(&self) -> RusotoFuture<IsVpcPeeredResult, IsVpcPeeredError>;

    fn open_instance_public_ports(
        &self, 
        input: OpenInstancePublicPortsRequest
    ) -> RusotoFuture<OpenInstancePublicPortsResult, OpenInstancePublicPortsError>;

    fn peer_vpc(&self) -> RusotoFuture<PeerVpcResult, PeerVpcError>;

    fn put_instance_public_ports(
        &self, 
        input: PutInstancePublicPortsRequest
    ) -> RusotoFuture<PutInstancePublicPortsResult, PutInstancePublicPortsError>;

    fn reboot_instance(
        &self, 
        input: RebootInstanceRequest
    ) -> RusotoFuture<RebootInstanceResult, RebootInstanceError>;

    fn release_static_ip(
        &self, 
        input: ReleaseStaticIpRequest
    ) -> RusotoFuture<ReleaseStaticIpResult, ReleaseStaticIpError>;

    fn start_instance(
        &self, 
        input: StartInstanceRequest
    ) -> RusotoFuture<StartInstanceResult, StartInstanceError>;

    fn stop_instance(
        &self, 
        input: StopInstanceRequest
    ) -> RusotoFuture<StopInstanceResult, StopInstanceError>;

    fn unpeer_vpc(&self) -> RusotoFuture<UnpeerVpcResult, UnpeerVpcError>;

    fn update_domain_entry(
        &self, 
        input: UpdateDomainEntryRequest
    ) -> RusotoFuture<UpdateDomainEntryResult, UpdateDomainEntryError>;

    fn update_load_balancer_attribute(
        &self, 
        input: UpdateLoadBalancerAttributeRequest
    ) -> RusotoFuture<UpdateLoadBalancerAttributeResult, UpdateLoadBalancerAttributeError>;
}

Trait representing the capabilities of the Amazon Lightsail API. Amazon Lightsail clients implement this trait.

Required Methods

fn allocate_static_ip(
&self,
input: AllocateStaticIpRequest
) -> RusotoFuture<AllocateStaticIpResult, AllocateStaticIpError>

Allocates a static IP address.

fn attach_disk(
&self,
input: AttachDiskRequest
) -> RusotoFuture<AttachDiskResult, AttachDiskError>

Attaches a block storage disk to a running or stopped Lightsail instance and exposes it to the instance with the specified disk name.

fn attach_instances_to_load_balancer(
&self,
input: AttachInstancesToLoadBalancerRequest
) -> RusotoFuture<AttachInstancesToLoadBalancerResult, AttachInstancesToLoadBalancerError>

Attaches one or more Lightsail instances to a load balancer.

After some time, the instances are attached to the load balancer and the health check status is available.

fn attach_load_balancer_tls_certificate(
&self,
input: AttachLoadBalancerTlsCertificateRequest
) -> RusotoFuture<AttachLoadBalancerTlsCertificateResult, AttachLoadBalancerTlsCertificateError>

Attaches a Transport Layer Security (TLS) certificate to your load balancer. TLS is just an updated, more secure version of Secure Socket Layer (SSL).

Once you create and validate your certificate, you can attach it to your load balancer. You can also use this API to rotate the certificates on your account. Use the AttachLoadBalancerTlsCertificate operation with the non-attached certificate, and it will replace the existing one and become the attached certificate.

fn attach_static_ip(
&self,
input: AttachStaticIpRequest
) -> RusotoFuture<AttachStaticIpResult, AttachStaticIpError>

Attaches a static IP address to a specific Amazon Lightsail instance.

fn close_instance_public_ports(
&self,
input: CloseInstancePublicPortsRequest
) -> RusotoFuture<CloseInstancePublicPortsResult, CloseInstancePublicPortsError>

Closes the public ports on a specific Amazon Lightsail instance.

fn create_disk(
&self,
input: CreateDiskRequest
) -> RusotoFuture<CreateDiskResult, CreateDiskError>

Creates a block storage disk that can be attached to a Lightsail instance in the same Availability Zone (e.g., us-east-2a). The disk is created in the regional endpoint that you send the HTTP request to. For more information, see Regions and Availability Zones in Lightsail.

fn create_disk_from_snapshot(
&self,
input: CreateDiskFromSnapshotRequest
) -> RusotoFuture<CreateDiskFromSnapshotResult, CreateDiskFromSnapshotError>

Creates a block storage disk from a disk snapshot that can be attached to a Lightsail instance in the same Availability Zone (e.g., us-east-2a). The disk is created in the regional endpoint that you send the HTTP request to. For more information, see Regions and Availability Zones in Lightsail.

fn create_disk_snapshot(
&self,
input: CreateDiskSnapshotRequest
) -> RusotoFuture<CreateDiskSnapshotResult, CreateDiskSnapshotError>

Creates a snapshot of a block storage disk. You can use snapshots for backups, to make copies of disks, and to save data before shutting down a Lightsail instance.

You can take a snapshot of an attached disk that is in use; however, snapshots only capture data that has been written to your disk at the time the snapshot command is issued. This may exclude any data that has been cached by any applications or the operating system. If you can pause any file systems on the disk long enough to take a snapshot, your snapshot should be complete. Nevertheless, if you cannot pause all file writes to the disk, you should unmount the disk from within the Lightsail instance, issue the create disk snapshot command, and then remount the disk to ensure a consistent and complete snapshot. You may remount and use your disk while the snapshot status is pending.

fn create_domain(
&self,
input: CreateDomainRequest
) -> RusotoFuture<CreateDomainResult, CreateDomainError>

Creates a domain resource for the specified domain (e.g., example.com).

fn create_domain_entry(
&self,
input: CreateDomainEntryRequest
) -> RusotoFuture<CreateDomainEntryResult, CreateDomainEntryError>

Creates one of the following entry records associated with the domain: A record, CNAME record, TXT record, or MX record.

fn create_instance_snapshot(
&self,
input: CreateInstanceSnapshotRequest
) -> RusotoFuture<CreateInstanceSnapshotResult, CreateInstanceSnapshotError>

Creates a snapshot of a specific virtual private server, or instance. You can use a snapshot to create a new instance that is based on that snapshot.

fn create_instances(
&self,
input: CreateInstancesRequest
) -> RusotoFuture<CreateInstancesResult, CreateInstancesError>

Creates one or more Amazon Lightsail virtual private servers, or instances.

fn create_instances_from_snapshot(
&self,
input: CreateInstancesFromSnapshotRequest
) -> RusotoFuture<CreateInstancesFromSnapshotResult, CreateInstancesFromSnapshotError>

Uses a specific snapshot as a blueprint for creating one or more new instances that are based on that identical configuration.

fn create_key_pair(
&self,
input: CreateKeyPairRequest
) -> RusotoFuture<CreateKeyPairResult, CreateKeyPairError>

Creates sn SSH key pair.

fn create_load_balancer(
&self,
input: CreateLoadBalancerRequest
) -> RusotoFuture<CreateLoadBalancerResult, CreateLoadBalancerError>

Creates a Lightsail load balancer. To learn more about deciding whether to load balance your application, see Configure your Lightsail instances for load balancing. You can create up to 5 load balancers per AWS Region in your account.

When you create a load balancer, you can specify a unique name and port settings. To change additional load balancer settings, use the UpdateLoadBalancerAttribute operation.

fn create_load_balancer_tls_certificate(
&self,
input: CreateLoadBalancerTlsCertificateRequest
) -> RusotoFuture<CreateLoadBalancerTlsCertificateResult, CreateLoadBalancerTlsCertificateError>

Creates a Lightsail load balancer TLS certificate.

TLS is just an updated, more secure version of Secure Socket Layer (SSL).

fn delete_disk(
&self,
input: DeleteDiskRequest
) -> RusotoFuture<DeleteDiskResult, DeleteDiskError>

Deletes the specified block storage disk. The disk must be in the available state (not attached to a Lightsail instance).

The disk may remain in the deleting state for several minutes.

fn delete_disk_snapshot(
&self,
input: DeleteDiskSnapshotRequest
) -> RusotoFuture<DeleteDiskSnapshotResult, DeleteDiskSnapshotError>

Deletes the specified disk snapshot.

When you make periodic snapshots of a disk, the snapshots are incremental, and only the blocks on the device that have changed since your last snapshot are saved in the new snapshot. When you delete a snapshot, only the data not needed for any other snapshot is removed. So regardless of which prior snapshots have been deleted, all active snapshots will have access to all the information needed to restore the disk.

fn delete_domain(
&self,
input: DeleteDomainRequest
) -> RusotoFuture<DeleteDomainResult, DeleteDomainError>

Deletes the specified domain recordset and all of its domain records.

fn delete_domain_entry(
&self,
input: DeleteDomainEntryRequest
) -> RusotoFuture<DeleteDomainEntryResult, DeleteDomainEntryError>

Deletes a specific domain entry.

fn delete_instance(
&self,
input: DeleteInstanceRequest
) -> RusotoFuture<DeleteInstanceResult, DeleteInstanceError>

Deletes a specific Amazon Lightsail virtual private server, or instance.

fn delete_instance_snapshot(
&self,
input: DeleteInstanceSnapshotRequest
) -> RusotoFuture<DeleteInstanceSnapshotResult, DeleteInstanceSnapshotError>

Deletes a specific snapshot of a virtual private server (or instance).

fn delete_key_pair(
&self,
input: DeleteKeyPairRequest
) -> RusotoFuture<DeleteKeyPairResult, DeleteKeyPairError>

Deletes a specific SSH key pair.

fn delete_load_balancer(
&self,
input: DeleteLoadBalancerRequest
) -> RusotoFuture<DeleteLoadBalancerResult, DeleteLoadBalancerError>

Deletes a Lightsail load balancer and all its associated SSL/TLS certificates. Once the load balancer is deleted, you will need to create a new load balancer, create a new certificate, and verify domain ownership again.

fn delete_load_balancer_tls_certificate(
&self,
input: DeleteLoadBalancerTlsCertificateRequest
) -> RusotoFuture<DeleteLoadBalancerTlsCertificateResult, DeleteLoadBalancerTlsCertificateError>

Deletes an SSL/TLS certificate associated with a Lightsail load balancer.

fn detach_disk(
&self,
input: DetachDiskRequest
) -> RusotoFuture<DetachDiskResult, DetachDiskError>

Detaches a stopped block storage disk from a Lightsail instance. Make sure to unmount any file systems on the device within your operating system before stopping the instance and detaching the disk.

fn detach_instances_from_load_balancer(
&self,
input: DetachInstancesFromLoadBalancerRequest
) -> RusotoFuture<DetachInstancesFromLoadBalancerResult, DetachInstancesFromLoadBalancerError>

Detaches the specified instances from a Lightsail load balancer.

This operation waits until the instances are no longer needed before they are detached from the load balancer.

fn detach_static_ip(
&self,
input: DetachStaticIpRequest
) -> RusotoFuture<DetachStaticIpResult, DetachStaticIpError>

Detaches a static IP from the Amazon Lightsail instance to which it is attached.

fn download_default_key_pair(
&self
) -> RusotoFuture<DownloadDefaultKeyPairResult, DownloadDefaultKeyPairError>

Downloads the default SSH key pair from the user's account.

fn get_active_names(
&self,
input: GetActiveNamesRequest
) -> RusotoFuture<GetActiveNamesResult, GetActiveNamesError>

Returns the names of all active (not deleted) resources.

fn get_blueprints(
&self,
input: GetBlueprintsRequest
) -> RusotoFuture<GetBlueprintsResult, GetBlueprintsError>

Returns the list of available instance images, or blueprints. You can use a blueprint to create a new virtual private server already running a specific operating system, as well as a preinstalled app or development stack. The software each instance is running depends on the blueprint image you choose.

fn get_bundles(
&self,
input: GetBundlesRequest
) -> RusotoFuture<GetBundlesResult, GetBundlesError>

Returns the list of bundles that are available for purchase. A bundle describes the specs for your virtual private server (or instance).

fn get_disk(
&self,
input: GetDiskRequest
) -> RusotoFuture<GetDiskResult, GetDiskError>

Returns information about a specific block storage disk.

fn get_disk_snapshot(
&self,
input: GetDiskSnapshotRequest
) -> RusotoFuture<GetDiskSnapshotResult, GetDiskSnapshotError>

Returns information about a specific block storage disk snapshot.

fn get_disk_snapshots(
&self,
input: GetDiskSnapshotsRequest
) -> RusotoFuture<GetDiskSnapshotsResult, GetDiskSnapshotsError>

Returns information about all block storage disk snapshots in your AWS account and region.

If you are describing a long list of disk snapshots, you can paginate the output to make the list more manageable. You can use the pageToken and nextPageToken values to retrieve the next items in the list.

fn get_disks(
&self,
input: GetDisksRequest
) -> RusotoFuture<GetDisksResult, GetDisksError>

Returns information about all block storage disks in your AWS account and region.

If you are describing a long list of disks, you can paginate the output to make the list more manageable. You can use the pageToken and nextPageToken values to retrieve the next items in the list.

fn get_domain(
&self,
input: GetDomainRequest
) -> RusotoFuture<GetDomainResult, GetDomainError>

Returns information about a specific domain recordset.

fn get_domains(
&self,
input: GetDomainsRequest
) -> RusotoFuture<GetDomainsResult, GetDomainsError>

Returns a list of all domains in the user's account.

fn get_instance(
&self,
input: GetInstanceRequest
) -> RusotoFuture<GetInstanceResult, GetInstanceError>

Returns information about a specific Amazon Lightsail instance, which is a virtual private server.

fn get_instance_access_details(
&self,
input: GetInstanceAccessDetailsRequest
) -> RusotoFuture<GetInstanceAccessDetailsResult, GetInstanceAccessDetailsError>

Returns temporary SSH keys you can use to connect to a specific virtual private server, or instance.

fn get_instance_metric_data(
&self,
input: GetInstanceMetricDataRequest
) -> RusotoFuture<GetInstanceMetricDataResult, GetInstanceMetricDataError>

Returns the data points for the specified Amazon Lightsail instance metric, given an instance name.

fn get_instance_port_states(
&self,
input: GetInstancePortStatesRequest
) -> RusotoFuture<GetInstancePortStatesResult, GetInstancePortStatesError>

Returns the port states for a specific virtual private server, or instance.

fn get_instance_snapshot(
&self,
input: GetInstanceSnapshotRequest
) -> RusotoFuture<GetInstanceSnapshotResult, GetInstanceSnapshotError>

Returns information about a specific instance snapshot.

fn get_instance_snapshots(
&self,
input: GetInstanceSnapshotsRequest
) -> RusotoFuture<GetInstanceSnapshotsResult, GetInstanceSnapshotsError>

Returns all instance snapshots for the user's account.

fn get_instance_state(
&self,
input: GetInstanceStateRequest
) -> RusotoFuture<GetInstanceStateResult, GetInstanceStateError>

Returns the state of a specific instance. Works on one instance at a time.

fn get_instances(
&self,
input: GetInstancesRequest
) -> RusotoFuture<GetInstancesResult, GetInstancesError>

Returns information about all Amazon Lightsail virtual private servers, or instances.

fn get_key_pair(
&self,
input: GetKeyPairRequest
) -> RusotoFuture<GetKeyPairResult, GetKeyPairError>

Returns information about a specific key pair.

fn get_key_pairs(
&self,
input: GetKeyPairsRequest
) -> RusotoFuture<GetKeyPairsResult, GetKeyPairsError>

Returns information about all key pairs in the user's account.

fn get_load_balancer(
&self,
input: GetLoadBalancerRequest
) -> RusotoFuture<GetLoadBalancerResult, GetLoadBalancerError>

Returns information about the specified Lightsail load balancer.

fn get_load_balancer_metric_data(
&self,
input: GetLoadBalancerMetricDataRequest
) -> RusotoFuture<GetLoadBalancerMetricDataResult, GetLoadBalancerMetricDataError>

Returns information about health metrics for your Lightsail load balancer.

fn get_load_balancer_tls_certificates(
&self,
input: GetLoadBalancerTlsCertificatesRequest
) -> RusotoFuture<GetLoadBalancerTlsCertificatesResult, GetLoadBalancerTlsCertificatesError>

Returns information about the TLS certificates that are associated with the specified Lightsail load balancer.

TLS is just an updated, more secure version of Secure Socket Layer (SSL).

You can have a maximum of 2 certificates associated with a Lightsail load balancer. One is active and the other is inactive.

fn get_load_balancers(
&self,
input: GetLoadBalancersRequest
) -> RusotoFuture<GetLoadBalancersResult, GetLoadBalancersError>

Returns information about all load balancers in an account.

If you are describing a long list of load balancers, you can paginate the output to make the list more manageable. You can use the pageToken and nextPageToken values to retrieve the next items in the list.

fn get_operation(
&self,
input: GetOperationRequest
) -> RusotoFuture<GetOperationResult, GetOperationError>

Returns information about a specific operation. Operations include events such as when you create an instance, allocate a static IP, attach a static IP, and so on.

fn get_operations(
&self,
input: GetOperationsRequest
) -> RusotoFuture<GetOperationsResult, GetOperationsError>

Returns information about all operations.

Results are returned from oldest to newest, up to a maximum of 200. Results can be paged by making each subsequent call to GetOperations use the maximum (last) statusChangedAt value from the previous request.

fn get_operations_for_resource(
&self,
input: GetOperationsForResourceRequest
) -> RusotoFuture<GetOperationsForResourceResult, GetOperationsForResourceError>

Gets operations for a specific resource (e.g., an instance or a static IP).

fn get_regions(
&self,
input: GetRegionsRequest
) -> RusotoFuture<GetRegionsResult, GetRegionsError>

Returns a list of all valid regions for Amazon Lightsail. Use the include availability zones parameter to also return the availability zones in a region.

fn get_static_ip(
&self,
input: GetStaticIpRequest
) -> RusotoFuture<GetStaticIpResult, GetStaticIpError>

Returns information about a specific static IP.

fn get_static_ips(
&self,
input: GetStaticIpsRequest
) -> RusotoFuture<GetStaticIpsResult, GetStaticIpsError>

Returns information about all static IPs in the user's account.

fn import_key_pair(
&self,
input: ImportKeyPairRequest
) -> RusotoFuture<ImportKeyPairResult, ImportKeyPairError>

Imports a public SSH key from a specific key pair.

fn is_vpc_peered(&self) -> RusotoFuture<IsVpcPeeredResult, IsVpcPeeredError>

Returns a Boolean value indicating whether your Lightsail VPC is peered.

fn open_instance_public_ports(
&self,
input: OpenInstancePublicPortsRequest
) -> RusotoFuture<OpenInstancePublicPortsResult, OpenInstancePublicPortsError>

Adds public ports to an Amazon Lightsail instance.

fn peer_vpc(&self) -> RusotoFuture<PeerVpcResult, PeerVpcError>

Tries to peer the Lightsail VPC with the user's default VPC.

fn put_instance_public_ports(
&self,
input: PutInstancePublicPortsRequest
) -> RusotoFuture<PutInstancePublicPortsResult, PutInstancePublicPortsError>

Sets the specified open ports for an Amazon Lightsail instance, and closes all ports for every protocol not included in the current request.

fn reboot_instance(
&self,
input: RebootInstanceRequest
) -> RusotoFuture<RebootInstanceResult, RebootInstanceError>

Restarts a specific instance. When your Amazon Lightsail instance is finished rebooting, Lightsail assigns a new public IP address. To use the same IP address after restarting, create a static IP address and attach it to the instance.

fn release_static_ip(
&self,
input: ReleaseStaticIpRequest
) -> RusotoFuture<ReleaseStaticIpResult, ReleaseStaticIpError>

Deletes a specific static IP from your account.

fn start_instance(
&self,
input: StartInstanceRequest
) -> RusotoFuture<StartInstanceResult, StartInstanceError>

Starts a specific Amazon Lightsail instance from a stopped state. To restart an instance, use the reboot instance operation.

fn stop_instance(
&self,
input: StopInstanceRequest
) -> RusotoFuture<StopInstanceResult, StopInstanceError>

Stops a specific Amazon Lightsail instance that is currently running.

fn unpeer_vpc(&self) -> RusotoFuture<UnpeerVpcResult, UnpeerVpcError>

Attempts to unpeer the Lightsail VPC from the user's default VPC.

fn update_domain_entry(
&self,
input: UpdateDomainEntryRequest
) -> RusotoFuture<UpdateDomainEntryResult, UpdateDomainEntryError>

Updates a domain recordset after it is created.

fn update_load_balancer_attribute(
&self,
input: UpdateLoadBalancerAttributeRequest
) -> RusotoFuture<UpdateLoadBalancerAttributeResult, UpdateLoadBalancerAttributeError>

Updates the specified attribute for a load balancer. You can only update one attribute at a time.